Split the law-automation module into cohesive submodules (no behavior change)
The law-automation auto-proof module had grown past 7.9k lines, mixing the
keystone laws-as-lemmas composition, the rational-over-floor sign/magnitude
family, and the structural-induction path in one file. Section-8 proof work
will keep editing it, so carve out two cohesive submodules:

  induction/keystone.rs    -- pool-composition recognizer/emitter, signed
                              power-of-two homomorphism, order-bridge citations,
                              and the shared arithmetic/expression helpers
  induction/floor_bound.rs -- the rational-over-floor sign/magnitude family
                              (Lemma 7.2.2 sign half) and its rf_* helpers
  induction/mod.rs         -- everything else (the inductive list-recursion path
                              and shared support), re-exporting the keystone
                              entry points so external call sites are unchanged

Pure module restructuring: only item moves, module declarations, use paths and
visibility adjustments. No tactic string, recognizer condition, dispatch order
or emitted Lean changed. The generated Lean is byte-identical for every K5
module (rational, estimate, fprep, round, table, kernel), and lib/proof_spec/
clippy all stay green.
