+++ src/codegen/lean/law_auto/induction/mod.rs
@@ -18,6 +18,13 @@ use super::shared::law_simp_defs;
 use crate::ast::{TypeDef, TypeVariant, VerifyBlock, VerifyLaw};
 use crate::codegen::CodegenContext;
 
+mod floor_bound;
+mod keystone;
+
+pub(in crate::codegen::lean) use keystone::{
+    emit_pool_composition_generic_law, recognize_pool_composition_generic,
+};
+
 /// `first | (<simp>; done) | (<simp>; omega) | sorry` — the per-arm portfolio
 /// shared by the Peano bridge theorems (`<simp>` is e.g. `simp [f]` /
 /// `simp [f, ih]`). A structured `First` so `--minimize` collapses the arm to
@@ -1961,3293 +1968,6 @@ pub(in crate::codegen::lean) fn emit_conditional_inductive_generic_law(
     Some(AutoProof {
         support_lines: support,
         body: Tactic::raw(body),
-        replaces_theorem: false,
-    })
-}
-
-/// Keystone — the NON-recursive analog of [`recognize_conditional_inductive_generic`].
-/// A conditional claim with NO list to induct on, decomposed purely by citing
-/// the laws-as-lemmas pool: EARLIER sibling laws in the file abstract whatever
-/// recursion the cone hides (a `pow2` homomorphism, a `mulLeMonoRight`
-/// monotonicity), so `grind` over those pool laws + the cone closes the goal
-/// WITHOUT inducting — grind e-matches each pool law, instantiates it, discharges
-/// its premise from context, and finishes the residual ring / linear identity.
-/// This is The Method's composition step made generic: the algebraic content
-/// lives as Aver helper laws (proven on their own), and the engine only supplies
-/// the citation skeleton, never a per-figure template.
-///
-/// Speculative: the probe is the oracle (grind either closes from the pool or it
-/// does not), and `default` is OFF — a no-probe transpile keeps the bounded
-/// sampled statement, so a law `grind`+pool cannot close never regresses off its
-/// sound bounded fallback.
-pub(in crate::codegen::lean) fn recognize_pool_composition_generic(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> bool {
-    // A non-conditional law is admitted ONLY for the equational power-of-two
-    // homomorphism composition: an earlier-sibling homomorphism law about a
-    // recursive power-of-two cone fn (detected name-blind via its `FloorDivWindow`
-    // figure) cited through the pow2 normalizer — the SIGNED pow2 homomorphism
-    // `2^(m+n) = 2^m·2^n` for ANY integers, whose sign branches each reduce to a
-    // product of `pow2` of nonneg atoms the normalizer canonicalizes. Every other
-    // unconditional shape keeps its own strategy (RingIdentity, LinearArithmetic,
-    // SimpOverPreludeLemmas, …): this can never steal one, because admission here
-    // requires both a power-of-two normalizer fn AND a non-empty equational pool,
-    // which those shapes do not have. (`when`-laws keep the full keystone surface —
-    // order-bridge, rational-floor and no-pool definitional arms below.)
-    if law.when.is_none()
-        && (keystone_pow2_fn(vb, law, ctx).is_none()
-            || keystone_pool_names(vb, law, ctx).is_empty())
-    {
-        return false;
-    }
-    // The inductive path owns list-recursion laws; this is the residue with no
-    // list-induction target.
-    if find_list_induction_target(law).is_some() {
-        return false;
-    }
-    // Exclusive with the comparison-bridge arm (it runs first).
-    if recognize_conditional_comparison_bridge(law, ctx) {
-        return false;
-    }
-    // The claim goes through the subject fn (`holds`, or an equational `=> rhs`).
-    if !matches!(&law.lhs.node, crate::ast::Expr::FnCall(..)) {
-        return false;
-    }
-    // Fail closed on refinement-lifted givens. A `@Nat>=0`-style refined given
-    // takes a DIFFERENT statement path: the `conditional_universal` gate requires
-    // `lifted_vars.is_empty()` (verify.rs), so a lifted law's universal statement
-    // carries the domain through the lifted quantifier / extra premises — which the
-    // keystone's fixed `intro <givens> h_when` proof shape does not match (the intro
-    // count would disagree with the statement's binders). Mirroring that gate here
-    // keeps the recognizer's contract symmetric across EVERY call site (statement
-    // and emit), so a lifted law can never trigger a keystone proof out of step with
-    // its statement. (No corpus law hits this today; this is a fail-closed guard.)
-    if law.givens.iter().any(|g| {
-        crate::codegen::common::refinement_lift_for_given(
-            &g.name,
-            &g.type_name,
-            &law.lhs,
-            &law.rhs,
-            ctx,
-        )
-        .is_some()
-    }) {
-        return false;
-    }
-    // This rung earns its keep over the (when-gated-off) plain simp / grind
-    // rungs in two ways, both probe-decided:
-    //   * a citable POOL — an earlier law whose subject fn is in this law's
-    //     call cone (the EQUATIONAL pool, e.g. a `pow2` homomorphism abstracting
-    //     cone recursion) or an earlier inequality `L <op> R` matching subterms
-    //     of the goal / premises (the ORDER-BRIDGE pool, e.g. a `mulLeMonoRight`
-    //     monotonicity feeding a transitivity step); OR
-    //   * NO pool at all, but the conditional goal is a flat definitional
-    //     identity `grind` closes outright once the cone is unfolded — a
-    //     record-field projection through `trunc` / `fpScale`, a Bool predicate
-    //     that reduces to `rfl` under the premises. The plain grind rung NEVER
-    //     runs on a `when`-law (it is gated to `law.when.is_none()`), so without
-    //     this arm such a law has no universal path and silently degrades to the
-    //     bounded sampled statement.
-    // The speculative probe is the oracle in BOTH cases: a candidate that does
-    // not close falls back to bounded (default `false`), so admitting the
-    // no-pool case is fail-safe and leaves a non-probe `transpile` byte-identical.
-    //
-    // The no-pool case is gated to laws WITHOUT a dedicated template strategy of
-    // their own. A `when`-law pinned to a bespoke template (the `FloorDivWindow`
-    // pow / window family — its own PowSumSplit / SigWindow / ProductWindow
-    // proofs need functional induction the keystone's bare `grind` can't do) is
-    // dispatched LATER and must keep that template; admitting it here would let
-    // the probe steal it with a grind that can only fail. Pool-citing laws skip
-    // this gate entirely (they are never `FloorDivWindow`-shaped and the pool is
-    // the whole point).
-    let has_pool = !keystone_pool_names(vb, law, ctx).is_empty()
-        || !keystone_order_bridge_citations(vb, law, ctx).is_empty();
-    // The rational-over-floor sign/magnitude family (Lemma 7.2.2) cites its
-    // earlier window/magnitude helper laws through its OWN dedicated arm
-    // (`rf_citations`), not the equational/order-bridge pools, so admit it here
-    // regardless of those pools or any catch-all strategy pin.
-    let rational_floor = rational_floor_shape(vb, law, ctx).is_some();
-    if !has_pool && !rational_floor {
-        use crate::ir::ProofStrategy;
-        match super::law_strategy_for(ctx, &vb.fn_name, &law.name) {
-            None
-            | Some(ProofStrategy::BackendDispatch)
-            | Some(ProofStrategy::LinearArithmetic { .. }) => {}
-            _ => return false,
-        }
-    }
-    let id = format!("{}.{}", vb.fn_name, law.name);
-    super::super::tactic_ir::speculative::admits(&id, false)
-}
-
-/// The laws-as-lemmas pool for the keystone: the Lean theorem NAMES of EARLIER
-/// sibling laws whose SUBJECT fn lies in this law's call cone. Looser than
-/// [`earlier_law_lemmas`] (the inductive simp path) on purpose — the keystone's
-/// whole point is to cite a law ABOUT a recursive cone fn (a `pow2`
-/// homomorphism) to abstract that recursion, which the strict pure-fn-subset
-/// eligibility excludes. `grind` instantiates and discharges each cited law;
-/// an irrelevant one is simply unused, so subject-in-cone is the right gate.
-///
-/// Pool membership is decided DETERMINISTICALLY by `law_as_lemma_statement`
-/// (static IR / source-order checks), NOT by the speculative `admits` tracking
-/// that gates the keystone law itself — so the cited set is identical in the
-/// probe build and the committed build. Soundness of citing a pool member does
-/// not rest on this stability anyway: if a cited law's own proof carries `sorry`,
-/// the citing theorem inherits `sorryAx` and the `--check` whitelist refuses it
-/// credit (see the `pinned_when_universal` note in `law_as_lemma_statement`).
-fn keystone_pool_names(vb: &VerifyBlock, law: &VerifyLaw, ctx: &CodegenContext) -> Vec<String> {
-    use crate::ast::{TopLevel, VerifyKind};
-    let inputs = crate::codegen::proof_lower::ProofLowerInputs::from_ctx(ctx);
-    let cone = crate::codegen::proof_lower::LawProofCone::compute(law, &vb.fn_name, &inputs);
-    let cone_fns: std::collections::HashSet<String> =
-        cone.pure_fns().iter().map(|fd| fd.name.clone()).collect();
-    let mut out = Vec::new();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        // Only blocks EARLIER in source — source order is emit order, so the
-        // cited theorem precedes this one and cyclic use is impossible.
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break;
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !cone_fns.contains(&prev.fn_name) {
-            continue;
-        }
-        if let Some((name, _stmt)) =
-            crate::codegen::lean::toplevel::law_as_lemma_statement(prev, prev_law, ctx)
-        {
-            out.push(name);
-        }
-    }
-    out
-}
-
-/// The SUBJECT fns (Lean names) of the keystone's cited equational pool —
-/// the fn each cited pool law is a rewrite ABOUT. Mirrors
-/// [`keystone_pool_names`] one-for-one but returns the subject fn instead
-/// of the lemma name. The keystone keeps these fns FOLDED (out of its
-/// `simp only` cone unfold) so the cited rewrite can e-match the goal term:
-/// the floorDiv exact-cancel `floor(s·pow(b), pow(a)) = s·pow(b−a)` can only
-/// fire while `floorDiv` is still abstract, exactly as the `pow2`
-/// homomorphism needs `pow2` abstract (recursive fns are folded for the same
-/// reason). Shape-keyed and name-blind — driven by which laws are cited, not
-/// by any hard-coded fn name.
-fn keystone_pool_subject_fns(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> Vec<String> {
-    use crate::ast::{TopLevel, VerifyKind};
-    let inputs = crate::codegen::proof_lower::ProofLowerInputs::from_ctx(ctx);
-    let cone = crate::codegen::proof_lower::LawProofCone::compute(law, &vb.fn_name, &inputs);
-    let cone_fns: std::collections::HashSet<String> =
-        cone.pure_fns().iter().map(|fd| fd.name.clone()).collect();
-    let mut out = Vec::new();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break;
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !cone_fns.contains(&prev.fn_name) {
-            continue;
-        }
-        if crate::codegen::lean::toplevel::law_as_lemma_statement(prev, prev_law, ctx).is_some() {
-            out.push(aver_name_to_lean(&prev.fn_name));
-        }
-    }
-    out
-}
-
-/// The cited pool laws (by Lean theorem name) that REGISTER THEMSELVES as
-/// `@[grind]` via a `grind_pattern` — the `FloorPow2Cancel` exact-division
-/// cancel, keyed on its floor term. They are already in `grind`'s lemma set
-/// globally, so re-listing them in `grind [...]` only earns a "redundant
-/// parameter" warning. The keystone drops them from its explicit hint list
-/// (but still keeps their subject fn FOLDED, via [`keystone_pool_subject_fns`],
-/// so their floor term survives the cone unfold for the pattern to fire).
-fn keystone_self_registering_pool_names(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> std::collections::HashSet<String> {
-    use crate::ast::{TopLevel, VerifyKind};
-    use crate::ir::{FloorWindowFigure, ProofStrategy};
-    let inputs = crate::codegen::proof_lower::ProofLowerInputs::from_ctx(ctx);
-    let cone = crate::codegen::proof_lower::LawProofCone::compute(law, &vb.fn_name, &inputs);
-    let cone_fns: std::collections::HashSet<String> =
-        cone.pure_fns().iter().map(|fd| fd.name.clone()).collect();
-    let mut out = std::collections::HashSet::new();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break;
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !cone_fns.contains(&prev.fn_name) {
-            continue;
-        }
-        if !matches!(
-            super::law_strategy_for(ctx, &prev.fn_name, &prev_law.name),
-            Some(ProofStrategy::FloorDivWindow {
-                figure: FloorWindowFigure::FloorPow2Cancel { .. }
-            })
-        ) {
-            continue;
-        }
-        if let Some((name, _stmt)) =
-            crate::codegen::lean::toplevel::law_as_lemma_statement(prev, prev_law, ctx)
-        {
-            out.insert(name);
-        }
-    }
-    out
-}
-
-/// The power-of-two function the keystone's cited pool reasons about, if any:
-/// the `pow_fn` of an EARLIER sibling law (in this law's cone) whose proof
-/// strategy is a `FloorDivWindow` figure (the power-of-two homomorphism /
-/// window family), returned in Lean form. `None` when no cited pool law is
-/// about a power-of-two fn — the keystone then keeps its plain simp+grind
-/// emission unchanged.
-///
-/// Shape-keyed and name-blind: the trigger is the `FloorDivWindow` figure
-/// (which the lowerer only pins after its `is_pow2_shape` gate), never a
-/// hard-coded fn name. When present, the keystone prepends the generic
-/// [`pow2_linear_form_normalizer_support`](super::floor_window) stack so
-/// `grind` can canonicalize `pow` of ANY nonneg linear form.
-fn keystone_pow2_fn(vb: &VerifyBlock, law: &VerifyLaw, ctx: &CodegenContext) -> Option<String> {
-    use crate::ast::{TopLevel, VerifyKind};
-    use crate::ir::{FloorWindowFigure, ProofStrategy};
-    let inputs = crate::codegen::proof_lower::ProofLowerInputs::from_ctx(ctx);
-    let cone = crate::codegen::proof_lower::LawProofCone::compute(law, &vb.fn_name, &inputs);
-    let cone_fns: std::collections::HashSet<String> =
-        cone.pure_fns().iter().map(|fd| fd.name.clone()).collect();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        // Only earlier-in-source siblings — same gate as `keystone_pool_names`.
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break;
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !cone_fns.contains(&prev.fn_name) {
-            continue;
-        }
-        let Some(ProofStrategy::FloorDivWindow { figure }) =
-            super::law_strategy_for(ctx, &prev.fn_name, &prev_law.name)
-        else {
-            continue;
-        };
-        let pow_fn = match figure {
-            FloorWindowFigure::PowPositive { pow_fn }
-            | FloorWindowFigure::PowSumSplit { pow_fn }
-            | FloorWindowFigure::SigWindow { pow_fn, .. }
-            | FloorWindowFigure::ProductWindow { pow_fn, .. }
-            | FloorWindowFigure::FloorPow2Window { pow_fn, .. }
-            | FloorWindowFigure::FloorPow2Cancel { pow_fn, .. } => pow_fn,
-        };
-        return Some(aver_name_to_lean(&pow_fn));
-    }
-    None
-}
-
-/// A SIGNED-power-of-two homomorphism citation for the keystone: when an
-/// earlier sibling law proves the homomorphism of a `Fraction`-valued
-/// signed power-of-two cone fn (`2^k` faithful for every integer `k`),
-/// and that fn is in this law's cone, the keystone can canonicalize a
-/// cross-multiplied product of it through the proven law — the `Fraction`
-/// analog of [`keystone_pow2_fn`]. All Lean names; `cited` is the proven
-/// homomorphism law theorem to cite.
-///
-/// Shape-keyed and name-blind: the trigger is (a) an earlier `… holds` law
-/// whose inlined body is `sameValue(F(m+n), times(F(m), F(n)))` for a
-/// single-arg fn `F`, (b) `F` returning `Fraction` with a `k < 0` sign
-/// match (the signed-power-of-two shape), and (c) `F`'s underlying integer
-/// power-of-two being exactly the one [`keystone_pow2_fn`] already
-/// canonicalizes (so both normalizers are present and consistent). No
-/// hard-coded fn or law name.
-struct Pow2SignedCite {
-    sgn: String,
-    pow: String,
-    hom_fn: String,
-    cited: String,
-}
-
-fn keystone_pow2signed_cite(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> Option<Pow2SignedCite> {
-    use crate::ast::{Expr, Literal, TopLevel, VerifyKind};
-    let int_pow = keystone_pow2_fn(vb, law, ctx)?;
-    let inputs = crate::codegen::proof_lower::ProofLowerInputs::from_ctx(ctx);
-    let cone = crate::codegen::proof_lower::LawProofCone::compute(law, &vb.fn_name, &inputs);
-    let cone_fns: std::collections::HashSet<String> =
-        cone.pure_fns().iter().map(|fd| fd.name.clone()).collect();
-    let all_fns: Vec<&crate::ast::FnDef> = inputs.pure_fns();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break;
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !matches!(&prev_law.rhs.node, Expr::Literal(Literal::Bool(true))) {
-            continue;
-        }
-        let Some(inlined) = inline_fn_call(&prev_law.lhs, ctx) else {
-            continue;
-        };
-        let Some(sgn_dotted) = match_signed_homomorphism_shape(&inlined.node) else {
-            continue;
-        };
-        let sgn_basename = sgn_dotted
-            .rsplit('.')
-            .next()
-            .unwrap_or(&sgn_dotted)
-            .to_string();
-        if !cone_fns.contains(&sgn_basename) {
-            continue;
-        }
-        let Some(sgn_fd) = all_fns.iter().copied().find(|fd| fd.name == sgn_basename) else {
-            continue;
-        };
-        if !signed_pow2_shape(sgn_fd, &int_pow) {
-            continue;
-        }
-        // The signed homomorphism only earns its keep when the goal applies `F`
-        // to a SUM of exponents — produced by a cone fn that builds a record with
-        // an ADDITIVE field (e.g. a float product's `exp = a.exp + b.exp`). A law
-        // that applies `F` only to a bare variable (`F(f.exp)`, the
-        // nonneg-exponent agreement) needs `F` UNFOLDED and reduced by its sign
-        // premise, not folded + canonicalized — folding it there would strand the
-        // single `F(var)` term. Gate on that additive-record shape (name-blind):
-        // no cone fn building a sum-valued field ⇒ no `F(sum)` ⇒ decline.
-        if !cone_has_additive_record(&cone_fns, &sgn_basename, &prev.fn_name, &all_fns) {
-            continue;
-        }
-        let Some((thm_name, _stmt)) =
-            crate::codegen::lean::toplevel::law_as_lemma_statement(prev, prev_law, ctx)
-        else {
-            continue;
-        };
-        return Some(Pow2SignedCite {
-            sgn: aver_name_to_lean(&sgn_basename),
-            pow: int_pow,
-            hom_fn: aver_name_to_lean(&prev.fn_name),
-            cited: thm_name,
-        });
-    }
-    None
-}
-
-/// Match the signed-power-of-two homomorphism body shape
-/// `sameValue(F(A + B), times(F(A), F(B)))`, returning `F`'s dotted name.
-/// `sameValue` / `times` are matched by basename (the Rational ops), `F`
-/// by structural consistency across the three calls — name-blind in `F`.
-fn match_signed_homomorphism_shape(e: &crate::ast::Expr) -> Option<String> {
-    use crate::ast::{BinOp, Expr};
-    let Expr::FnCall(callee, args) = e else {
-        return None;
-    };
-    if args.len() != 2 {
-        return None;
-    }
-    let cname = super::shared::expr_dotted_name(callee)?;
-    if cname.rsplit('.').next() != Some("sameValue") {
-        return None;
-    }
-    let Expr::FnCall(f0, f0args) = &args[0].node else {
-        return None;
-    };
-    if f0args.len() != 1 {
-        return None;
-    }
-    let f0name = super::shared::expr_dotted_name(f0)?;
-    let Expr::BinOp(BinOp::Add, a, b) = &f0args[0].node else {
-        return None;
-    };
-    let Expr::FnCall(tcallee, targs) = &args[1].node else {
-        return None;
-    };
-    if targs.len() != 2 {
-        return None;
-    }
-    let tname = super::shared::expr_dotted_name(tcallee)?;
-    if tname.rsplit('.').next() != Some("times") {
-        return None;
-    }
-    let Expr::FnCall(fa, faargs) = &targs[0].node else {
-        return None;
-    };
-    let Expr::FnCall(fb, fbargs) = &targs[1].node else {
-        return None;
-    };
-    if faargs.len() != 1 || fbargs.len() != 1 {
-        return None;
-    }
-    if super::shared::expr_dotted_name(fa)? != f0name
-        || super::shared::expr_dotted_name(fb)? != f0name
-    {
-        return None;
-    }
-    if faargs[0].node != a.node || fbargs[0].node != b.node {
-        return None;
-    }
-    Some(f0name)
-}
-
-/// `F` is a signed-power-of-two fn: one `Int` param, returns `Fraction`,
-/// body is a `k < 0` sign match, and it calls the SAME integer
-/// power-of-two `int_pow` (Lean name) the keystone already canonicalizes.
-fn signed_pow2_shape(fd: &crate::ast::FnDef, int_pow: &str) -> bool {
-    use crate::ast::{BinOp, Expr, Stmt};
-    let [(p, ty)] = fd.params.as_slice() else {
-        return false;
-    };
-    if ty != "Int" || fd.return_type.rsplit('.').next() != Some("Fraction") {
-        return false;
-    }
-    let [Stmt::Expr(body)] = fd.body.stmts() else {
-        return false;
-    };
-    let Expr::Match { subject, arms } = &body.node else {
-        return false;
-    };
-    let Expr::BinOp(BinOp::Lt, sl, sr) = &subject.node else {
-        return false;
-    };
-    if super::shared::expr_dotted_name(sl).as_deref() != Some(p)
-        || !matches!(&sr.node, Expr::Literal(crate::ast::Literal::Int(0)))
-    {
-        return false;
-    }
-    if arms.len() != 2 {
-        return false;
-    }
-    // The body must canonicalize through the SAME integer power-of-two the
-    // keystone's pow2 normalizer uses — collect every fn the arms call and
-    // require `int_pow` among them (name-blind: matched by Lean name equality
-    // with the already-detected `int_pow`, never a literal).
-    let mut called: Vec<String> = Vec::new();
-    for arm in arms {
-        collect_fncall_names(&arm.body.node, &mut called);
-    }
-    // Compare on the BARE basename: a cross-module consumer holds `int_pow`
-    // qualified (`Domain.Fprep.pow2`) while the signed fn's own body calls the
-    // pow fn by its in-module bare name (`pow2`) — match the tails so the
-    // recognizer fires for a dep-module signed power of two too.
-    let want = rf_bare_basename(int_pow);
-    called
-        .iter()
-        .any(|n| rf_bare_basename(&aver_name_to_lean(n)) == want)
-}
-
-/// Collect the dotted names of every `FnCall` in `e` (recursively).
-fn collect_fncall_names(e: &crate::ast::Expr, out: &mut Vec<String>) {
-    use crate::ast::Expr;
-    match e {
-        Expr::FnCall(callee, args) => {
-            if let Some(n) = super::shared::expr_dotted_name(callee) {
-                out.push(n);
-            }
-            for a in args {
-                collect_fncall_names(&a.node, out);
-            }
-        }
-        Expr::BinOp(_, a, b) => {
-            collect_fncall_names(&a.node, out);
-            collect_fncall_names(&b.node, out);
-        }
-        Expr::Neg(a) => collect_fncall_names(&a.node, out),
-        Expr::Attr(b, _) => collect_fncall_names(&b.node, out),
-        Expr::RecordCreate { fields, .. } => {
-            for (_, v) in fields {
-                collect_fncall_names(&v.node, out);
-            }
-        }
-        Expr::Match { subject, arms } => {
-            collect_fncall_names(&subject.node, out);
-            for arm in arms {
-                collect_fncall_names(&arm.body.node, out);
-            }
-        }
-        _ => {}
-    }
-}
-
-/// True when some fn in `cone_fns` (other than the signed fn itself or the
-/// homomorphism fn) builds a record with at least one field whose value is a
-/// `BinOp::Add` — the static signature of a fn that produces a sum-valued
-/// exponent the signed power-of-two is then applied to (e.g. a float product's
-/// `exp = a.exp + b.exp`). Name-blind: keyed on the `RecordCreate` + `Add`
-/// shape, never on a fn name.
-fn cone_has_additive_record(
-    cone_fns: &std::collections::HashSet<String>,
-    sgn_basename: &str,
-    hom_basename: &str,
-    all_fns: &[&crate::ast::FnDef],
-) -> bool {
-    for fd in all_fns {
-        if !cone_fns.contains(&fd.name) || fd.name == sgn_basename || fd.name == hom_basename {
-            continue;
-        }
-        for stmt in fd.body.stmts() {
-            if let crate::ast::Stmt::Expr(e) = stmt
-                && expr_has_additive_record_field(&e.node)
-            {
-                return true;
-            }
-        }
-    }
-    false
-}
-
-/// Recursively: does `e` contain a `RecordCreate` with a field whose value is a
-/// `BinOp::Add`?
-fn expr_has_additive_record_field(e: &crate::ast::Expr) -> bool {
-    use crate::ast::Expr;
-    match e {
-        Expr::RecordCreate { fields, .. } => fields.iter().any(|(_, v)| {
-            matches!(&v.node, Expr::BinOp(crate::ast::BinOp::Add, _, _))
-                || expr_has_additive_record_field(&v.node)
-        }),
-        Expr::FnCall(c, args) => {
-            expr_has_additive_record_field(&c.node)
-                || args.iter().any(|a| expr_has_additive_record_field(&a.node))
-        }
-        Expr::BinOp(_, a, b) => {
-            expr_has_additive_record_field(&a.node) || expr_has_additive_record_field(&b.node)
-        }
-        Expr::Neg(a) => expr_has_additive_record_field(&a.node),
-        Expr::Attr(b, _) => expr_has_additive_record_field(&b.node),
-        Expr::Match { subject, arms } => {
-            expr_has_additive_record_field(&subject.node)
-                || arms
-                    .iter()
-                    .any(|arm| expr_has_additive_record_field(&arm.body.node))
-        }
-        Expr::RecordUpdate { base, updates, .. } => {
-            updates.iter().any(|(_, v)| {
-                matches!(&v.node, Expr::BinOp(crate::ast::BinOp::Add, _, _))
-                    || expr_has_additive_record_field(&v.node)
-            }) || expr_has_additive_record_field(&base.node)
-        }
-        _ => false,
-    }
-}
-
-/// One eligible ORDER-BRIDGE citation for the keystone: an EARLIER sibling law
-/// asserting an inequality `L <op> R` (its subject fn's body is a single
-/// comparison, both sides non-atomic, `… holds`) whose two sides match subterms
-/// of the citing law's goal / premises under one consistent substitution of the
-/// cited law's givens. `name` is the Prop-form RESTATEMENT theorem (a `grind`
-/// hint for the citing law); `support` is that theorem's declaration plus its
-/// `grind_pattern` over the two comparison sides. The restatement is DERIVED
-/// from the cited law's Bool theorem (`simpa … using (<thm> …)`), so deleting
-/// the Aver law breaks it — a genuine citation, not a re-proof.
-struct OrderBridgeCitation {
-    name: String,
-    support: Vec<String>,
-}
-
-/// The Lean infix for a comparison `BinOp`, or `None` for a non-comparison op
-/// (an order law's body must be one of these four).
-fn comparison_op_lean(op: crate::ast::BinOp) -> Option<&'static str> {
-    use crate::ast::BinOp;
-    match op {
-        BinOp::Lt => Some("<"),
-        BinOp::Gt => Some(">"),
-        BinOp::Lte => Some("<="),
-        BinOp::Gte => Some(">="),
-        _ => None,
-    }
-}
-
-/// The variable NAME of a leaf, treating a source `Ident` and a resolver-pass
-/// `Resolved` slot uniformly — function bodies are emitted in resolved form
-/// (`Resolved`), while law claims/premises are source form (`Ident`), and the
-/// order-bridge matches one against the other.
-fn expr_var_name(e: &crate::ast::Expr) -> Option<&str> {
-    match e {
-        crate::ast::Expr::Ident(n) => Some(n),
-        crate::ast::Expr::Resolved { name, .. } => Some(name),
-        _ => None,
-    }
-}
-
-/// An atomic term (bare variable or literal) is neither a valid `grind_pattern`
-/// trigger nor a meaningful comparison side, so an order law with an atomic side
-/// is rejected.
-fn expr_is_atomic(e: &crate::ast::Expr) -> bool {
-    expr_var_name(e).is_some() || matches!(e, crate::ast::Expr::Literal(_))
-}
-
-/// Structural equality on arithmetic terms, span-insensitive and identifying a
-/// source `Ident(n)` with a resolved `Resolved{name:n}` (used to re-bind a
-/// matched pattern variable consistently).
-fn arith_eq(a: &crate::ast::Expr, b: &crate::ast::Expr) -> bool {
-    use crate::ast::Expr;
-    match (expr_var_name(a), expr_var_name(b)) {
-        (Some(na), Some(nb)) => return na == nb,
-        (None, None) => {}
-        _ => return false,
-    }
-    match (a, b) {
-        (Expr::Literal(x), Expr::Literal(y)) => x == y,
-        (Expr::BinOp(o1, a1, b1), Expr::BinOp(o2, a2, b2)) => {
-            o1 == o2 && arith_eq(&a1.node, &a2.node) && arith_eq(&b1.node, &b2.node)
-        }
-        (Expr::Neg(x), Expr::Neg(y)) => arith_eq(&x.node, &y.node),
-        (Expr::FnCall(c1, a1), Expr::FnCall(c2, a2)) => {
-            a1.len() == a2.len()
-                && arith_eq(&c1.node, &c2.node)
-                && a1.iter().zip(a2).all(|(p, q)| arith_eq(&p.node, &q.node))
-        }
-        (Expr::Attr(b1, f1), Expr::Attr(b2, f2)) => f1 == f2 && arith_eq(&b1.node, &b2.node),
-        _ => false,
-    }
-}
-
-/// Whether `name` occurs as a free variable anywhere in the arithmetic term.
-fn ident_occurs(e: &crate::ast::Expr, name: &str) -> bool {
-    use crate::ast::Expr;
-    if expr_var_name(e) == Some(name) {
-        return true;
-    }
-    match e {
-        Expr::BinOp(_, a, b) => ident_occurs(&a.node, name) || ident_occurs(&b.node, name),
-        Expr::Neg(a) => ident_occurs(&a.node, name),
-        Expr::FnCall(c, args) => {
-            ident_occurs(&c.node, name) || args.iter().any(|x| ident_occurs(&x.node, name))
-        }
-        Expr::Attr(b, _) => ident_occurs(&b.node, name),
-        _ => false,
-    }
-}
-
-/// Push every arithmetic subterm of `expr` (itself included) onto `out`, by
-/// value. Walks the product/sum/comparison/call/negation/field skeleton; opaque
-/// leaves (matches, constructors, …) stop the recursion.
-fn collect_arith_subterms(
-    expr: &crate::ast::Spanned<crate::ast::Expr>,
-    out: &mut Vec<crate::ast::Expr>,
-) {
-    use crate::ast::Expr;
-    out.push(expr.node.clone());
-    match &expr.node {
-        Expr::BinOp(_, a, b) => {
-            collect_arith_subterms(a, out);
-            collect_arith_subterms(b, out);
-        }
-        Expr::Neg(a) => collect_arith_subterms(a, out),
-        Expr::FnCall(c, args) => {
-            collect_arith_subterms(c, out);
-            for a in args {
-                collect_arith_subterms(a, out);
-            }
-        }
-        Expr::Attr(b, _) => collect_arith_subterms(b, out),
-        _ => {}
-    }
-}
-
-/// Deep-clone `e`, replacing any free `Ident(n)` with `map[n]` (substitute a
-/// function's parameters by a call's argument terms). `Spanned::bare` is fine —
-/// the result feeds the Lean expr emitter, which ignores spans.
-fn substitute_idents(
-    e: &crate::ast::Spanned<crate::ast::Expr>,
-    map: &std::collections::HashMap<String, crate::ast::Spanned<crate::ast::Expr>>,
-) -> crate::ast::Spanned<crate::ast::Expr> {
-    use crate::ast::{Expr, Spanned};
-    let node = match &e.node {
-        Expr::Ident(n) => {
-            if let Some(rep) = map.get(n) {
-                return rep.clone();
-            }
-            Expr::Ident(n.clone())
-        }
-        // Function bodies are emitted in resolved form, so a parameter reference
-        // is a `Resolved` slot, not a source `Ident`. Substitute it by name too.
-        Expr::Resolved { name, .. } => {
-            if let Some(rep) = map.get(name) {
-                return rep.clone();
-            }
-            e.node.clone()
-        }
-        Expr::BinOp(op, a, b) => Expr::BinOp(
-            *op,
-            Box::new(substitute_idents(a, map)),
-            Box::new(substitute_idents(b, map)),
-        ),
-        Expr::Neg(a) => Expr::Neg(Box::new(substitute_idents(a, map))),
-        Expr::FnCall(c, args) => Expr::FnCall(
-            Box::new(substitute_idents(c, map)),
-            args.iter().map(|x| substitute_idents(x, map)).collect(),
-        ),
-        Expr::Attr(b, f) => Expr::Attr(Box::new(substitute_idents(b, map)), f.clone()),
-        // Record construction / update: a value-level fn (`fpValue`, `fpTrunc`)
-        // whose body builds a record must have its FIELD expressions substituted
-        // too, or a parameter referenced only inside a field (e.g. `fpTrunc`'s
-        // `n` inside `sigBits = floorDiv(.. pow2(n-1) ..)`) would survive
-        // un-renamed — the catch-all `other.clone()` did not recurse here.
-        Expr::RecordCreate { type_name, fields } => Expr::RecordCreate {
-            type_name: type_name.clone(),
-            fields: fields
-                .iter()
-                .map(|(fname, v)| (fname.clone(), substitute_idents(v, map)))
-                .collect(),
-        },
-        Expr::RecordUpdate {
-            type_name,
-            base,
-            updates,
-        } => Expr::RecordUpdate {
-            type_name: type_name.clone(),
-            base: Box::new(substitute_idents(base, map)),
-            updates: updates
-                .iter()
-                .map(|(fname, v)| (fname.clone(), substitute_idents(v, map)))
-                .collect(),
-        },
-        Expr::Match { subject, arms } => Expr::Match {
-            subject: Box::new(substitute_idents(subject, map)),
-            arms: arms
-                .iter()
-                .map(|arm| crate::ast::MatchArm {
-                    body: Box::new(substitute_idents(&arm.body, map)),
-                    ..arm.clone()
-                })
-                .collect(),
-        },
-        other => other.clone(),
-    };
-    Spanned::bare(node)
-}
-
-/// Inline a `fn(args)` call: substitute the callee's parameters by `args` in its
-/// (single-expression) body. Returns the inlined body — for an order law this is
-/// the comparison `L <op> R`; for the citing law it is the goal arithmetic. Bails
-/// on anything that is not a direct call to a known single-tail function with a
-/// matching arity.
-fn inline_fn_call(
-    call: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-) -> Option<crate::ast::Spanned<crate::ast::Expr>> {
-    use crate::ast::Expr;
-    let Expr::FnCall(callee, args) = &call.node else {
-        return None;
-    };
-    let name = super::shared::expr_dotted_name(callee)?;
-    let fd = ctx.fn_def_by_name(&name, ctx.active_module_scope().as_deref())?;
-    if fd.params.len() != args.len() {
-        return None;
-    }
-    let body = fd.body.tail_expr()?;
-    let mut map: std::collections::HashMap<String, crate::ast::Spanned<crate::ast::Expr>> =
-        std::collections::HashMap::new();
-    for ((pname, _), arg) in fd.params.iter().zip(args.iter()) {
-        map.insert(pname.clone(), arg.clone());
-    }
-    Some(substitute_idents(body, &map))
-}
-
-/// First-order matching: does `pat` (over pattern variables `vars`) match the
-/// ground term `term`, extending `subst` consistently? Re-binding a variable
-/// checks `arith_eq` (span-insensitive, `Ident`/`Resolved`-agnostic).
-fn match_arith_pattern(
-    pat: &crate::ast::Expr,
-    term: &crate::ast::Expr,
-    vars: &std::collections::HashSet<String>,
-    subst: &mut std::collections::HashMap<String, crate::ast::Expr>,
-) -> bool {
-    use crate::ast::Expr;
-    // Leaf variable (source `Ident` or resolved slot): a pattern variable binds /
-    // re-checks; a non-variable leaf must be the same variable.
-    if let Some(pname) = expr_var_name(pat) {
-        if vars.contains(pname) {
-            return match subst.get(pname) {
-                Some(prev) => arith_eq(prev, term),
-                None => {
-                    subst.insert(pname.to_string(), term.clone());
-                    true
-                }
-            };
-        }
-        return expr_var_name(term) == Some(pname);
-    }
-    match pat {
-        Expr::Literal(l) => matches!(term, Expr::Literal(tl) if tl == l),
-        Expr::Neg(a) => {
-            if let Expr::Neg(ta) = term {
-                match_arith_pattern(&a.node, &ta.node, vars, subst)
-            } else {
-                false
-            }
-        }
-        Expr::BinOp(op, a, b) => {
-            if let Expr::BinOp(top, ta, tb) = term {
-                op == top
-                    && match_arith_pattern(&a.node, &ta.node, vars, subst)
-                    && match_arith_pattern(&b.node, &tb.node, vars, subst)
-            } else {
-                false
-            }
-        }
-        Expr::FnCall(c, args) => {
-            if let Expr::FnCall(tc, targs) = term {
-                args.len() == targs.len()
-                    && match_arith_pattern(&c.node, &tc.node, vars, subst)
-                    && args
-                        .iter()
-                        .zip(targs.iter())
-                        .all(|(p, t)| match_arith_pattern(&p.node, &t.node, vars, subst))
-            } else {
-                false
-            }
-        }
-        Expr::Attr(b, f) => {
-            if let Expr::Attr(tb, tf) = term {
-                f == tf && match_arith_pattern(&b.node, &tb.node, vars, subst)
-            } else {
-                false
-            }
-        }
-        _ => false,
-    }
-}
-
-/// Whether both comparison sides match subterms of the citing law under ONE
-/// consistent substitution (a variable shared between the sides — `c` in
-/// `a*c`/`b*c` — must map to the same citing subterm on both).
-fn order_sides_match_subterms(
-    lhs_side: &crate::ast::Expr,
-    rhs_side: &crate::ast::Expr,
-    vars: &std::collections::HashSet<String>,
-    subterms: &[crate::ast::Expr],
-) -> bool {
-    for t1 in subterms {
-        let mut s = std::collections::HashMap::new();
-        if match_arith_pattern(lhs_side, t1, vars, &mut s) {
-            for t2 in subterms {
-                let mut s2 = s.clone();
-                if match_arith_pattern(rhs_side, t2, vars, &mut s2) {
-                    return true;
-                }
-            }
-        }
-    }
-    false
-}
-
-/// The keystone's ORDER-BRIDGE pool for `law`: every earlier sibling inequality
-/// law whose two comparison sides match subterms of this law's goal / premises,
-/// each carried as a Prop-form restatement + `grind_pattern` (see
-/// [`OrderBridgeCitation`]). Deterministic (static IR + source order), so the
-/// recognizer, the statement gate, and the emit all see the same set.
-fn keystone_order_bridge_citations(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> Vec<OrderBridgeCitation> {
-    use crate::ast::{Expr, Literal, Spanned, TopLevel, VerifyKind};
-
-    // Subterms of the citing law's GOAL (the inlined subject body) and PREMISES.
-    let mut subterms: Vec<Expr> = Vec::new();
-    if let Some(when) = &law.when {
-        collect_arith_subterms(when, &mut subterms);
-    }
-    if let Some(goal) = inline_fn_call(&law.lhs, ctx) {
-        collect_arith_subterms(&goal, &mut subterms);
-    }
-    collect_arith_subterms(&law.rhs, &mut subterms);
-    if subterms.is_empty() {
-        return Vec::new();
-    }
-
-    let citing_uid = format!(
-        "{}_{}",
-        aver_name_to_lean(&vb.fn_name),
-        aver_name_to_lean(&law.name)
-    );
-
-    let mut out = Vec::new();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        // Only EARLIER blocks — source order is emit order, so the cited theorem
-        // precedes this one and cyclic citation is impossible.
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break;
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        // The cited law must assert the inequality HOLDS (`… holds` / `=> true`),
-        // so its Bool theorem concludes `<fn> args = true` and the restatement is
-        // sound; and it must expose that universal theorem to derive from.
-        if !matches!(&prev_law.rhs.node, Expr::Literal(Literal::Bool(true))) {
-            continue;
-        }
-        let Some((thm_name, _stmt)) =
-            crate::codegen::lean::toplevel::law_as_lemma_statement(prev, prev_law, ctx)
-        else {
-            continue;
-        };
-        // ORDER-LAW SHAPE: the cited fn's body is a single comparison `L <op> R`,
-        // inlined at the law's call args, with BOTH sides non-atomic.
-        let Some(inlined) = inline_fn_call(&prev_law.lhs, ctx) else {
-            continue;
-        };
-        let Expr::BinOp(op, l_box, r_box) = &inlined.node else {
-            continue;
-        };
-        let Some(op_lean) = comparison_op_lean(*op) else {
-            continue;
-        };
-        let lhs_side = &l_box.node;
-        let rhs_side = &r_box.node;
-        if expr_is_atomic(lhs_side) || expr_is_atomic(rhs_side) {
-            continue;
-        }
-        let given_names: Vec<String> = prev_law.givens.iter().map(|g| g.name.clone()).collect();
-        if given_names.is_empty() {
-            continue;
-        }
-        // Every quantified given must occur in a comparison side, so the
-        // `grind_pattern` over the two sides covers all the restatement's binders
-        // (`grind_pattern` errors on an uncovered variable).
-        if !given_names
-            .iter()
-            .all(|g| ident_occurs(lhs_side, g) || ident_occurs(rhs_side, g))
-        {
-            continue;
-        }
-        let vars: std::collections::HashSet<String> = given_names.iter().cloned().collect();
-        if !order_sides_match_subterms(lhs_side, rhs_side, &vars, &subterms) {
-            continue;
-        }
-
-        // Build the Prop-form restatement DERIVED from the Bool theorem. The name
-        // carries neither `_law_` nor `_eq_`, so the `--check` audit never mistakes
-        // a restatement for a creditable main-law theorem.
-        let rname = format!(
-            "{}_order_{}_{}",
-            citing_uid,
-            aver_name_to_lean(&prev.fn_name),
-            aver_name_to_lean(&prev_law.name)
-        );
-        let cited_fn_lean = aver_name_to_lean(&prev.fn_name);
-        let quant = prev_law
-            .givens
-            .iter()
-            .map(|g| {
-                format!(
-                    "({} : {})",
-                    aver_name_to_lean(&g.name),
-                    crate::codegen::lean::types::type_annotation_to_lean(&g.type_name)
-                )
-            })
-            .collect::<Vec<_>>()
-            .join(" ");
-        let names_csv = given_names
-            .iter()
-            .map(|n| aver_name_to_lean(n))
-            .collect::<Vec<_>>()
-            .join(" ");
-        let l_span: &Spanned<Expr> = l_box;
-        let r_span: &Spanned<Expr> = r_box;
-        let l_lean = super::super::expr::emit_expr_legacy(l_span, ctx, None);
-        let r_lean = super::super::expr::emit_expr_legacy(r_span, ctx, None);
-        let (premise_clause, intro, apply_args) = match &prev_law.when {
-            Some(when) => {
-                let p = super::super::expr::emit_expr_legacy(when, ctx, None);
-                (
-                    format!("{p} = true -> "),
-                    format!("  intro {names_csv} hp"),
-                    format!("{names_csv} hp"),
-                )
-            }
-            None => (
-                String::new(),
-                format!("  intro {names_csv}"),
-                names_csv.clone(),
-            ),
-        };
-        let support = vec![
-            format!(
-                "theorem {rname} : ∀ {quant}, {premise_clause}({l_lean}) {op_lean} ({r_lean}) := by"
-            ),
-            intro,
-            format!(
-                "  simpa only [{cited_fn_lean}, decide_eq_true_eq] using ({thm_name} {apply_args})"
-            ),
-            format!("grind_pattern {rname} => ({l_lean}), ({r_lean})"),
-        ];
-        out.push(OrderBridgeCitation {
-            name: rname,
-            support,
-        });
-    }
-    out
-}
-
-// ===========================================================================
-// Rational-over-floor sign / magnitude family (Lemma 7.2.2, p.19 sign half).
-//
-// A `when`-law whose subject reduces to either a comparison-against-zero
-// (a nonnegativity / strict-positivity magnitude fact) or a `match` on a Bool
-// predicate (the sign split), AND whose call cone reaches a power-of-two fn.
-// The keystone arm here is a GENERIC, name-blind skeleton:
-//   * it CITES earlier sibling `holds` laws whose subject is a single
-//     order comparison or a `Bool.and` of comparisons (the floor window, the
-//     magnitude/positivity helper laws) over a prefix of this law's givens,
-//     discharging each cited law's premise by projecting this law's `when`;
-//   * it supplies power-of-two positivity `have`s (shape-keyed on the cone's
-//     `pow2` fn) at the record-field exponents, so the nonlinear closer can
-//     discharge the power-of-two factor leaves;
-//   * the nonneg/pos shape closes with `aver_int_order` (the same nonlinear
-//     primitive the Newton-Raphson bounds use); the sign shape `rcases` the
-//     predicate's leading disjunction (the float sign) and `grind`s each arm
-//     against the cited magnitude/positivity facts.
-// The whole arm sits inside the keystone's `first | … | sorry` floor, so any
-// citation/discharge that does not typecheck falls to the honest floor and the
-// law degrades to its bounded sampled statement — credit stays fail-closed.
-// The algebraic content (the magnitude product, the value-magnitude product,
-// the window bounds) lives ENTIRELY in the cited Aver laws; deleting any of
-// them breaks the proof.
-// ===========================================================================
-
-#[derive(Clone, Copy, PartialEq)]
-enum RationalFloorShape {
-    /// subject body is `E >= 0` / `0 <= E` / `E > 0` / `0 < E`.
-    NonnegPos,
-    /// subject body is `match <pred-call> { true -> …; false -> … }`.
-    Sign,
-    /// subject body is the rational-order truncation-error bound
-    /// `lessThan(absFraction(minus(fpValue(F), fpValue(G))), pow2Signed(LE))`
-    /// (Lemma 7.2.2's strict bound). `G` is the rounded value of `F`; the
-    /// bound arm factors the error magnitude, splits the `pow2Signed` sign
-    /// branch, CITES the floor window + power-of-two homomorphism pool laws,
-    /// and multiplies the cited window remainder bound by the positive product
-    /// factor (the generic multiply-by-positive rung).
-    Bound,
-}
-
-/// The power-of-two fn (Lean name) reachable in this law's cone, if any. Tries
-/// the keystone's cited-pool detector first (same-module homomorphism cone),
-/// then falls back to the `pow_fn` of ANY `FloorDivWindow`-figured law in the
-/// module — the floor-window family pins that figure only on an `is_pow2_shape`
-/// fn, so this is name-blind. Guarded to a `pow` that actually appears in this
-/// law's cone, so an unrelated module's `pow` never leaks in.
-fn rf_pow2_fn(vb: &VerifyBlock, law: &VerifyLaw, ctx: &CodegenContext) -> Option<String> {
-    use crate::ast::{TopLevel, VerifyKind};
-    use crate::ir::{FloorWindowFigure, ProofStrategy};
-    let cone = super::shared::law_simp_source_names(ctx, vb, law);
-    if let Some(p) = keystone_pow2_fn(vb, law, ctx) {
-        return Some(p);
-    }
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if let Some(ProofStrategy::FloorDivWindow { figure }) =
-            super::law_strategy_for(ctx, &prev.fn_name, &prev_law.name)
-        {
-            let pow = match figure {
-                FloorWindowFigure::PowPositive { pow_fn }
-                | FloorWindowFigure::PowSumSplit { pow_fn }
-                | FloorWindowFigure::SigWindow { pow_fn, .. }
-                | FloorWindowFigure::ProductWindow { pow_fn, .. }
-                | FloorWindowFigure::FloorPow2Window { pow_fn, .. }
-                | FloorWindowFigure::FloorPow2Cancel { pow_fn, .. } => pow_fn,
-            };
-            let pbare = rf_bare_basename(&pow);
-            if cone.iter().any(|c| rf_bare_basename(c) == pbare) {
-                return Some(aver_name_to_lean(&pow));
-            }
-        }
-    }
-    None
-}
-
-/// Recognize the rational-over-floor sign/magnitude family on `(vb, law)`.
-/// Name-blind: keyed on the inlined subject body shape plus the presence of a
-/// power-of-two fn in the cone (via [`rf_pow2_fn`]). Returns `None` (decline)
-/// for every other keystone law, so their emission stays byte-identical.
-fn rational_floor_shape(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> Option<RationalFloorShape> {
-    use crate::ast::{BinOp, Expr, Literal};
-    rf_pow2_fn(vb, law, ctx)?;
-    let inlined = inline_fn_call(&law.lhs, ctx)?;
-    let is_zero = |e: &Expr| matches!(e, Expr::Literal(Literal::Int(0)));
-    // The rational-order bound `lessThan(absFraction(_), pow2Signed(_))`: a
-    // `lessThan` whose left arg is an `absFraction` and right arg a `pow2Signed`.
-    // Name-blind (basename match), keyed on the call shape; the `Bound` arm
-    // below extracts the float / precision givens and the rounded value from it.
-    if let Expr::FnCall(callee, args) = &inlined.node
-        && rf_callee_basename(callee).as_deref() == Some("lessThan")
-        && args.len() == 2
-        && matches!(&args[0].node, Expr::FnCall(c, a)
-            if rf_callee_basename(c).as_deref() == Some("absFraction") && a.len() == 1)
-        && matches!(&args[1].node, Expr::FnCall(c, a)
-            if rf_callee_basename(c).as_deref() == Some("pow2Signed") && a.len() == 1)
-    {
-        return Some(RationalFloorShape::Bound);
-    }
-    match &inlined.node {
-        Expr::BinOp(BinOp::Gte | BinOp::Gt, _l, r) if is_zero(&r.node) => {
-            Some(RationalFloorShape::NonnegPos)
-        }
-        Expr::BinOp(BinOp::Lte | BinOp::Lt, l, _r) if is_zero(&l.node) => {
-            Some(RationalFloorShape::NonnegPos)
-        }
-        Expr::Match { subject, arms }
-            if matches!(&subject.node, Expr::FnCall(..)) && arms.len() == 2 =>
-        {
-            Some(RationalFloorShape::Sign)
-        }
-        _ => None,
-    }
-}
-
-/// Basename of a call's callee (last dotted component), or `None` if the callee
-/// is not a name.
-fn rf_callee_basename(callee: &crate::ast::Spanned<crate::ast::Expr>) -> Option<String> {
-    super::shared::expr_dotted_name(callee).map(|n| rf_bare_basename(&n).to_string())
-}
-
-/// Flatten a left-folded `Bool.and(a, b)` premise (how the parser composes
-/// multiple `when` lines) into its atomic conjuncts.
-fn rf_flatten_bool_and(
-    e: &crate::ast::Spanned<crate::ast::Expr>,
-) -> Vec<&crate::ast::Spanned<crate::ast::Expr>> {
-    use crate::ast::Expr;
-    if let Expr::FnCall(callee, args) = &e.node
-        && super::shared::expr_dotted_name(callee).as_deref() == Some("Bool.and")
-        && args.len() == 2
-    {
-        let mut out = rf_flatten_bool_and(&args[0]);
-        out.extend(rf_flatten_bool_and(&args[1]));
-        return out;
-    }
-    vec![e]
-}
-
-/// A Lean proof term for `<cited_when> = true`, projected out of this law's
-/// `h_when` (the citing law's composite premise). `None` when the cited premise
-/// is not one of the citing law's conjuncts (then the citation is skipped — the
-/// arm keeps its honest floor).
-fn rf_premise_proof(
-    citing_when: &crate::ast::Spanned<crate::ast::Expr>,
-    cited_when: &crate::ast::Spanned<crate::ast::Expr>,
-) -> Option<String> {
-    let conjs = rf_flatten_bool_and(citing_when);
-    let n = conjs.len();
-    let idx = conjs
-        .iter()
-        .position(|c| arith_eq(&c.node, &cited_when.node))?;
-    if n == 1 {
-        return Some("h_when".to_string());
-    }
-    // `simp only [Bool.and_eq_true] at h_when` turns `(c0 && … && c_{n-1}) = true`
-    // into the LEFT-nested `((c0=true ∧ c1=true) ∧ …) ∧ c_{n-1}=true`. The flat
-    // conjunct at `idx` is reached by `.1` down the left spine then `.2`
-    // (`c0` is the whole left spine, `.1`×(n-1)). A `by`-block keeps the `at
-    // h_when` mutation local to this premise's elaboration. The omega arms are a
-    // fallback for a numeric Bool conjunct whose standalone elaboration differs
-    // from the composed `decide (…)` form (projection then fails the type-check).
-    let path = if idx == 0 {
-        ".1".repeat(n - 1)
-    } else if idx == n - 1 {
-        ".2".to_string()
-    } else {
-        format!("{}.2", ".1".repeat(n - 1 - idx))
-    };
-    Some(format!(
-        "(by first | (simp only [Bool.and_eq_true] at h_when; exact h_when{path}) | (simp only [Bool.and_eq_true, decide_eq_true_eq] at h_when ⊢ <;> omega) | (simp only [Bool.and_eq_true] at h_when; omega))"
-    ))
-}
-
-/// Whether an expression's head is a `Bool.or` (directly, or as the first
-/// conjunct of a `Bool.and`) — the shape of a sign/format predicate's body
-/// (`isFp`: `(sign == 1 || sign == -1) && …`), so its first projection is the
-/// sign disjunction to `rcases`.
-fn rf_starts_with_or(e: &crate::ast::Expr) -> bool {
-    use crate::ast::Expr;
-    if let Expr::FnCall(callee, args) = e {
-        let name = super::shared::expr_dotted_name(callee);
-        if name.as_deref() == Some("Bool.or") {
-            return true;
-        }
-        if name.as_deref() == Some("Bool.and")
-            && let Some(first) = args.first()
-        {
-            return rf_starts_with_or(&first.node);
-        }
-    }
-    false
-}
-
-/// Resolve a (possibly module-qualified) fn name to its def, searching the entry
-/// module and every dependency module by basename — `ctx.fn_def_by_name` does
-/// not resolve a dep fn by its qualified call name in the consumer scope.
-fn rf_resolve_fn<'a>(ctx: &'a CodegenContext, dotted: &str) -> Option<&'a crate::ast::FnDef> {
-    if let Some(fd) = ctx.fn_def_by_name(dotted, ctx.active_module_scope().as_deref()) {
-        return Some(fd);
-    }
-    let bare = dotted.rsplit('.').next().unwrap_or(dotted);
-    ctx.fn_defs.iter().find(|fd| fd.name == bare).or_else(|| {
-        ctx.modules
-            .iter()
-            .flat_map(|m| m.fn_defs.iter())
-            .find(|fd| fd.name == bare)
-    })
-}
-
-/// The `when`-conjunct that is the sign/format predicate (a fn call whose body
-/// leads with a disjunction), if any — the conjunct `rcases` splits on.
-fn rf_sign_conjunct<'a>(
-    law: &'a VerifyLaw,
-    ctx: &CodegenContext,
-) -> Option<&'a crate::ast::Spanned<crate::ast::Expr>> {
-    use crate::ast::Expr;
-    let when = law.when.as_ref()?;
-    rf_flatten_bool_and(when).into_iter().find(|c| {
-        let Expr::FnCall(callee, _) = &c.node else {
-            return false;
-        };
-        super::shared::expr_dotted_name(callee)
-            .and_then(|name| rf_resolve_fn(ctx, &name))
-            .and_then(|fd| fd.body.tail_expr())
-            .is_some_and(|body| rf_starts_with_or(&body.node))
-    })
-}
-
-/// Whether the named fn is the Euclidean floor wrapper
-/// (`Result.withDefault(Int.div(…), <lit>)`) — excluded from the simp unfold
-/// set so the floor term stays an opaque atom shared between the goal and the
-/// cited window bound (the consistency the nonlinear closer relies on).
-fn rf_is_floordiv_wrapper(ctx: &CodegenContext, bare: &str) -> bool {
-    use crate::ast::Expr;
-    let Some(fd) = ctx.fn_def_by_name(bare, ctx.active_module_scope().as_deref()) else {
-        return false;
-    };
-    let Some(body) = fd.body.tail_expr() else {
-        return false;
-    };
-    let Expr::FnCall(callee, args) = &body.node else {
-        return false;
-    };
-    super::shared::expr_dotted_name(callee).as_deref() == Some("Result.withDefault")
-        && args.len() == 2
-}
-
-fn rf_bare_basename(n: &str) -> &str {
-    let n = n.strip_prefix("_root_.").unwrap_or(n);
-    n.rsplit('.').next().unwrap_or(n)
-}
-
-/// The cone simp set for a law, EXCLUDING the recursive power-of-two fn and the
-/// Euclidean floor wrapper (both must stay opaque atoms). Used both for the
-/// citing goal and for unfolding each cited hypothesis.
-fn rf_filtered_defs(ctx: &CodegenContext, vb: &VerifyBlock, law: &VerifyLaw) -> Vec<String> {
-    let recursive: std::collections::HashSet<String> = super::recursive_pure_fn_names(ctx)
-        .iter()
-        .map(|n| rf_bare_basename(&aver_name_to_lean(n)).to_string())
-        .collect();
-    law_simp_defs(ctx, vb, law)
-        .into_iter()
-        .filter(|d| {
-            let b = rf_bare_basename(d);
-            !recursive.contains(b) && !rf_is_floordiv_wrapper(ctx, b)
-        })
-        .collect()
-}
-
-/// Int-typed field names of a (possibly dep-module) record type.
-fn rf_record_int_fields(ctx: &CodegenContext, type_name: &str) -> Vec<String> {
-    use crate::ast::TypeDef;
-    let base = type_name.rsplit('.').next().unwrap_or(type_name).trim();
-    let find = |tds: &[TypeDef]| -> Option<Vec<String>> {
-        tds.iter().find_map(|td| match td {
-            TypeDef::Product { name, fields, .. } if name == base => Some(
-                fields
-                    .iter()
-                    .filter(|(_, ty)| ty.trim() == "Int")
-                    .map(|(n, _)| n.clone())
-                    .collect(),
-            ),
-            _ => None,
-        })
-    };
-    if let Some(v) = find(&ctx.type_defs) {
-        return v;
-    }
-    for m in &ctx.modules {
-        if let Some(v) = find(&m.type_defs) {
-            return v;
-        }
-    }
-    Vec::new()
-}
-
-/// One cited earlier law for the rational-floor arm.
-struct RfCitation {
-    have_name: String,
-    apply: String,
-    simp_set: String,
-}
-
-/// Discover the earlier sibling `holds` laws to cite: subject body is a single
-/// order comparison or a `Bool.and` of comparisons, givens are a prefix (same
-/// names and types, in order) of this law's givens, and any premise projects
-/// out of this law's `when`. Deterministic (static IR + source order).
-fn rf_citations(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Vec<RfCitation> {
-    use crate::ast::{BinOp, Expr, TopLevel, VerifyKind};
-    let mut out = Vec::new();
-    let citing_when = law.when.as_ref();
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break; // only earlier-in-source siblings
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !matches!(
-            &prev_law.rhs.node,
-            Expr::Literal(crate::ast::Literal::Bool(true))
-        ) {
-            continue;
-        }
-        // Givens must be a prefix of the citing law's givens (same name+type).
-        if prev_law.givens.len() > law.givens.len() {
-            continue;
-        }
-        if !prev_law
-            .givens
-            .iter()
-            .zip(law.givens.iter())
-            .all(|(a, b)| a.name == b.name && a.type_name.trim() == b.type_name.trim())
-        {
-            continue;
-        }
-        // Subject body: a single order comparison, or a `Bool.and` of comparisons.
-        let Some(inlined) = inline_fn_call(&prev_law.lhs, ctx) else {
-            continue;
-        };
-        let is_cmp = |e: &Expr| {
-            matches!(
-                e,
-                Expr::BinOp(BinOp::Lt | BinOp::Gt | BinOp::Lte | BinOp::Gte, _, _)
-            )
-        };
-        let is_order_content = match &inlined.node {
-            Expr::FnCall(callee, args)
-                if super::shared::expr_dotted_name(callee).as_deref() == Some("Bool.and")
-                    && args.len() == 2 =>
-            {
-                is_cmp(&args[0].node) && is_cmp(&args[1].node)
-            }
-            other => is_cmp(other),
-        };
-        if !is_order_content {
-            continue;
-        }
-        // Premise discharge (if any) must project out of this law's `when`.
-        let prem = match &prev_law.when {
-            None => String::new(),
-            Some(w) => {
-                let Some(citing) = citing_when else {
-                    continue;
-                };
-                match rf_premise_proof(citing, w) {
-                    Some(p) => format!(" {p}"),
-                    None => continue,
-                }
-            }
-        };
-        let thm = format!("{}_law_{}", aver_name_to_lean(&prev.fn_name), prev_law.name);
-        let args = intro_names[..prev_law.givens.len()].join(" ");
-        let mut simp: Vec<String> = rf_filtered_defs(ctx, prev, prev_law);
-        simp.extend(
-            [
-                "Bool.and_eq_true",
-                "decide_eq_true_eq",
-                "ge_iff_le",
-                "gt_iff_lt",
-            ]
-            .iter()
-            .map(|s| s.to_string()),
-        );
-        let have_name = format!("h_{}_{}", aver_name_to_lean(&prev.fn_name), prev_law.name);
-        out.push(RfCitation {
-            apply: format!("{thm} {args}{prem}"),
-            simp_set: simp.join(", "),
-            have_name,
-        });
-    }
-    out
-}
-
-/// Emit a Lean string for an inlined sub-expression (used to derive the
-/// float-format field accessors and the floor term from the recognized shape —
-/// produced from the SAME definitions the goal unfolds, so they parse-match).
-fn rf_emit(ctx: &CodegenContext, e: &crate::ast::Spanned<crate::ast::Expr>) -> String {
-    super::super::expr::emit_expr_legacy(e, ctx, None)
-}
-
-/// One-level inline of a fn call, resolving the callee across the entry module
-/// AND dependency modules (unlike [`inline_fn_call`], which only resolves entry
-/// scope — a cross-module `fpValue` call would not resolve there). Substitutes
-/// the call's args for the resolved fn's params in its tail body.
-fn rf_inline_fn_call(
-    call: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-) -> Option<crate::ast::Spanned<crate::ast::Expr>> {
-    use crate::ast::Expr;
-    let Expr::FnCall(callee, args) = &call.node else {
-        return None;
-    };
-    let name = super::shared::expr_dotted_name(callee)?;
-    let fd = rf_resolve_fn(ctx, &name)?;
-    if fd.params.len() != args.len() {
-        return None;
-    }
-    let body = fd.body.tail_expr()?;
-    let mut map: std::collections::HashMap<String, crate::ast::Spanned<crate::ast::Expr>> =
-        std::collections::HashMap::new();
-    for ((pname, _), arg) in fd.params.iter().zip(args.iter()) {
-        map.insert(pname.clone(), arg.clone());
-    }
-    Some(substitute_idents(body, &map))
-}
-
-/// The four float-format value accessors (`f.sign`, `f.sigBits`, `f.exp`,
-/// `f.width` as emitted Lean strings) read off the inlined `fpValue(F)` body
-/// `Fraction(top = F.sign * F.sigBits * pow(F.exp), bottom = pow(F.width - 1))`.
-/// Name-blind — derived from `fpValue`'s own definition, never hard-coded.
-struct RfFpFields {
-    sign: String,
-    sig: String,
-    exp: String,
-    width: String,
-}
-
-fn rf_fpvalue_fields(
-    frac: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-) -> Option<RfFpFields> {
-    use crate::ast::{BinOp, Expr};
-    let Expr::RecordCreate { fields, .. } = &frac.node else {
-        return None;
-    };
-    let top = fields.iter().find(|(n, _)| n == "top").map(|(_, e)| e)?;
-    let bottom = fields.iter().find(|(n, _)| n == "bottom").map(|(_, e)| e)?;
-    // bottom = pow(Sub(Attr(F, width), _))
-    let Expr::FnCall(_, bargs) = &bottom.node else {
-        return None;
-    };
-    let Expr::BinOp(BinOp::Sub, wbase, _) = &bargs.first()?.node else {
-        return None;
-    };
-    // top = Mul(Mul(Attr(F, sign), Attr(F, sig)), pow(Attr(F, exp)))
-    let Expr::BinOp(BinOp::Mul, l, r) = &top.node else {
-        return None;
-    };
-    let Expr::BinOp(BinOp::Mul, sign_b, sig_b) = &l.node else {
-        return None;
-    };
-    let Expr::FnCall(_, eargs) = &r.node else {
-        return None;
-    };
-    let exp_b = eargs.first()?;
-    Some(RfFpFields {
-        sign: rf_emit(ctx, sign_b),
-        sig: rf_emit(ctx, sig_b),
-        exp: rf_emit(ctx, exp_b),
-        width: rf_emit(ctx, wbase),
-    })
-}
-
-/// The earlier-in-source sibling FLOOR WINDOW law to cite: a `holds` law whose
-/// givens are a prefix of this law's and whose inlined body is a `Bool.and` of
-/// two comparisons (the Euclidean window `W*q <= N ∧ N < W*(q+1)`). Returns its
-/// Lean theorem name and the unfold target. Name-blind (keyed on the body
-/// shape), deterministic (source order).
-fn rf_window_law(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-) -> Option<(String, String)> {
-    use crate::ast::{BinOp, Expr, TopLevel, VerifyKind};
-    let is_cmp = |e: &Expr| {
-        matches!(
-            e,
-            Expr::BinOp(BinOp::Lt | BinOp::Gt | BinOp::Lte | BinOp::Gte, _, _)
-        )
-    };
-    for item in &ctx.items {
-        let TopLevel::Verify(prev) = item else {
-            continue;
-        };
-        if prev.line == vb.line && prev.fn_name == vb.fn_name {
-            break; // earlier-in-source only
-        }
-        let VerifyKind::Law(prev_law) = &prev.kind else {
-            continue;
-        };
-        if !matches!(
-            &prev_law.rhs.node,
-            Expr::Literal(crate::ast::Literal::Bool(true))
-        ) {
-            continue;
-        }
-        if prev_law.givens.len() > law.givens.len()
-            || !prev_law
-                .givens
-                .iter()
-                .zip(law.givens.iter())
-                .all(|(a, b)| a.name == b.name && a.type_name.trim() == b.type_name.trim())
-        {
-            continue;
-        }
-        let Some(inlined) = inline_fn_call(&prev_law.lhs, ctx) else {
-            continue;
-        };
-        let Expr::FnCall(callee, args) = &inlined.node else {
-            continue;
-        };
-        if super::shared::expr_dotted_name(callee).as_deref() != Some("Bool.and") || args.len() != 2
-        {
-            continue;
-        }
-        if !is_cmp(&args[0].node) || !is_cmp(&args[1].node) {
-            continue;
-        }
-        let thm = format!("{}_law_{}", aver_name_to_lean(&prev.fn_name), prev_law.name);
-        let unfold = format!("_root_.{}", aver_name_to_lean(&prev.fn_name));
-        return Some((thm, unfold));
-    }
-    None
-}
-
-/// The power-of-two HOMOMORPHISM pool law's Lean theorem name (`pow(m+n) =
-/// pow(m)*pow(n)`), found by shape over the in-file laws and dep-module law
-/// pools. Name-blind: matches any law whose subject is the `pow` fn and whose
-/// claim's rhs is a product of two `pow` calls. The dep-module name is qualified
-/// with the module prefix (`Domain.Fprep.pow2_law_homomorphism`).
-fn rf_homomorphism_name(ctx: &CodegenContext, pow_lean: &str) -> Option<String> {
-    use crate::ast::{BinOp, Expr, TopLevel, VerifyKind};
-    let pow_base = rf_bare_basename(pow_lean);
-    let is_hom = |lw: &VerifyLaw| -> bool {
-        // rhs is `pow(_) * pow(_)`.
-        if let Expr::BinOp(BinOp::Mul, a, b) = &lw.rhs.node {
-            let is_pow_call = |e: &Expr| {
-                matches!(e, Expr::FnCall(c, _)
-                    if super::shared::expr_dotted_name(c)
-                        .as_deref()
-                        .map(rf_bare_basename) == Some(pow_base))
-            };
-            return is_pow_call(&a.node) && is_pow_call(&b.node);
-        }
-        false
-    };
-    // In-file (entry module) first.
-    for item in &ctx.items {
-        if let TopLevel::Verify(prev) = item
-            && let VerifyKind::Law(prev_law) = &prev.kind
-            && rf_bare_basename(&prev.fn_name) == pow_base
-            && is_hom(prev_law)
-        {
-            return Some(format!(
-                "{}_law_{}",
-                aver_name_to_lean(&prev.fn_name),
-                prev_law.name
-            ));
-        }
-    }
-    // Dep modules — qualify with the module prefix.
-    for m in &ctx.modules {
-        for prev in &m.verify_laws {
-            if let VerifyKind::Law(prev_law) = &prev.kind
-                && rf_bare_basename(&prev.fn_name) == pow_base
-                && is_hom(prev_law)
-            {
-                return Some(format!(
-                    "{}.{}_law_{}",
-                    m.prefix,
-                    aver_name_to_lean(&prev.fn_name),
-                    prev_law.name
-                ));
-            }
-        }
-    }
-    None
-}
-
-/// The rational-order truncation-error bound arm (Lemma 7.2.2's strict bound
-/// `|epsilon| < 2^(e_x - i + 1)`). Mirrors the de-risked litmus-clean proof: it
-/// factors the error magnitude (the sign cancels, the floor remainder is the
-/// nonneg core), splits the `pow2Signed` ulp-exponent sign branch, CITES the
-/// floor window pool law for the strict remainder bound `r < W` and the
-/// power-of-two homomorphism pool law for the ulp/significand exponent link, and
-/// MULTIPLIES the window bound by the positive product factor via the generic
-/// multiply-by-positive rung in `aver_int_order`. The algebra lives ENTIRELY in
-/// the two cited Aver laws; the arm only factors / sign-splits / multiplies /
-/// ring-bridges (`grind`). Returns `None` (decline → bounded fallback) on any
-/// structural surprise — credit stays fail-closed behind the probe + axiom
-/// whitelist.
-/// The four general-exponent value accessors read off an inlined
-/// `fpValueGeneral(F)` body `times(pow2Signed(F.exp), Fraction(top = F.sign *
-/// F.sigBits, bottom = pow(F.width - 1)))`: the signed-power-of-two cone fn (the
-/// `times` factor whose call shape is [`signed_pow2_shape`]), its exponent
-/// argument, and the sign / significand / width off the inner `Fraction`.
-/// Name-blind — derived from the value's own structure, never hard-coded.
-struct RfGeneralFields {
-    sgn_fn: String,
-    exp: String,
-    sign: String,
-    sig: String,
-    width: String,
-}
-
-/// Read the general-exponent value fields off an inlined `fpValueGeneral(F)`
-/// (`times(SGN(exp), Fraction(top = sign * sig, bottom = pow(width - 1)))`).
-/// Returns `None` if the value is not in that signed-power-of-two × fraction
-/// form (e.g. the clamped `fpValue`, a bare `Fraction`).
-fn rf_general_value_fields(
-    val: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-    int_pow: &str,
-) -> Option<RfGeneralFields> {
-    use crate::ast::{BinOp, Expr};
-    // val = times(A, B)
-    let Expr::FnCall(tcallee, targs) = &val.node else {
-        return None;
-    };
-    if super::shared::expr_dotted_name(tcallee)?.rsplit('.').next() != Some("times")
-        || targs.len() != 2
-    {
-        return None;
-    }
-    // A = SGN(exp), with SGN a signed-power-of-two cone fn (shape-recognized).
-    let Expr::FnCall(acallee, aargs) = &targs[0].node else {
-        return None;
-    };
-    if aargs.len() != 1 {
-        return None;
-    }
-    let sgn_dotted = super::shared::expr_dotted_name(acallee)?;
-    let sgn_fd = rf_resolve_fn(ctx, &sgn_dotted)?;
-    if !signed_pow2_shape(sgn_fd, int_pow) {
-        return None;
-    }
-    // B = Fraction { top = sign * sig, bottom = pow(width - 1) }.
-    let Expr::RecordCreate { fields, .. } = &targs[1].node else {
-        return None;
-    };
-    let top = fields.iter().find(|(n, _)| n == "top").map(|(_, e)| e)?;
-    let bottom = fields.iter().find(|(n, _)| n == "bottom").map(|(_, e)| e)?;
-    let Expr::BinOp(BinOp::Mul, sign_b, sig_b) = &top.node else {
-        return None;
-    };
-    let Expr::FnCall(_, bargs) = &bottom.node else {
-        return None;
-    };
-    let Expr::BinOp(BinOp::Sub, wbase, _) = &bargs.first()?.node else {
-        return None;
-    };
-    Some(RfGeneralFields {
-        sgn_fn: aver_name_to_lean(&sgn_dotted),
-        exp: rf_emit(ctx, &aargs[0]),
-        sign: rf_emit(ctx, sign_b),
-        sig: rf_emit(ctx, sig_b),
-        width: rf_emit(ctx, wbase),
-    })
-}
-
-/// One leaf of a piecewise rounded value: its significand integer `sig` (the
-/// `sigBits` field) and its exponent `exp`, as inlined record fields. The bound
-/// rung emits one `away_leaf` application per DISTINCT leaf.
-struct RfLeaf {
-    sig: crate::ast::Spanned<crate::ast::Expr>,
-    exp: crate::ast::Spanned<crate::ast::Expr>,
-    width: crate::ast::Spanned<crate::ast::Expr>,
-}
-
-/// Expand a record `sigBits` expression into the significand integers it can
-/// take: if it is a fn call that inlines to a (record-free) `match` — `awaySig`'s
-/// exact/round-up split — return one per arm (recursively); otherwise the bare
-/// expression. The arms here are integer terms (`floorDiv …`, `floorDiv … + 1`),
-/// not records, so the recursion bottoms out immediately for `away`/`sticky`.
-fn rf_expand_sig(
-    sig: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-) -> Vec<crate::ast::Spanned<crate::ast::Expr>> {
-    use crate::ast::Expr;
-    // Only expand a significand SELECTOR (`awaySig`) — never the recursive
-    // power-of-two or the floor wrapper (both stay opaque atoms).
-    if let Expr::FnCall(c, _) = &sig.node
-        && let Some(name) = super::shared::expr_dotted_name(c)
-    {
-        let base = rf_bare_basename(&name);
-        let recursive: std::collections::HashSet<String> = super::recursive_pure_fn_names(ctx)
-            .iter()
-            .map(|n| rf_bare_basename(&aver_name_to_lean(n)).to_string())
-            .collect();
-        if !recursive.contains(base)
-            && !rf_is_floordiv_wrapper(ctx, base)
-            && let Some(inl) = rf_inline_fn_call(sig, ctx)
-            && let Expr::Match { arms, .. } = &inl.node
-        {
-            return arms
-                .iter()
-                .flat_map(|a| rf_expand_sig(&a.body, ctx))
-                .collect();
-        }
-    }
-    vec![sig.clone()]
-}
-
-/// Recursively inline every non-recursive, non-`floorDiv`-wrapper user-fn call in
-/// `e` — mirroring what the rung's `set1` simp does to the goal, so the leaf's
-/// significand / floor terms computed here parse-match the unfolded goal. The
-/// Euclidean `floorDiv` wrapper and the recursive power-of-two stay opaque (a
-/// round-to-odd leaf's `2 * stickyHalf(f, n)` thus becomes `2 * floorDiv(...)`,
-/// exposing the half-precision floor the window law is about).
-fn rf_inline_wrappers(
-    e: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-) -> crate::ast::Spanned<crate::ast::Expr> {
-    let recursive: std::collections::HashSet<String> = super::recursive_pure_fn_names(ctx)
-        .iter()
-        .map(|n| rf_bare_basename(&aver_name_to_lean(n)).to_string())
-        .collect();
-    rf_inline_wrappers_inner(e, ctx, &recursive)
-}
-
-fn rf_inline_wrappers_inner(
-    e: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-    recursive: &std::collections::HashSet<String>,
-) -> crate::ast::Spanned<crate::ast::Expr> {
-    use crate::ast::{Expr, Spanned};
-    let rec = |x: &crate::ast::Spanned<Expr>| rf_inline_wrappers_inner(x, ctx, recursive);
-    let mapped = match &e.node {
-        Expr::BinOp(op, a, b) => {
-            Spanned::bare(Expr::BinOp(*op, Box::new(rec(a)), Box::new(rec(b))))
-        }
-        Expr::Neg(a) => Spanned::bare(Expr::Neg(Box::new(rec(a)))),
-        Expr::FnCall(c, args) => {
-            Spanned::bare(Expr::FnCall(c.clone(), args.iter().map(rec).collect()))
-        }
-        _ => e.clone(),
-    };
-    if let Expr::FnCall(c, _) = &mapped.node
-        && let Some(name) = super::shared::expr_dotted_name(c)
-    {
-        let base = rf_bare_basename(&aver_name_to_lean(&name)).to_string();
-        if !recursive.contains(&base)
-            && !rf_is_floordiv_wrapper(ctx, &base)
-            && let Some(inl) = rf_inline_fn_call(&mapped, ctx)
-        {
-            return rf_inline_wrappers_inner(&inl, ctx, recursive);
-        }
-    }
-    mapped
-}
-
-/// The per-leaf floor placement of a PIECEWISE rounded value's significand, read
-/// off the goal-form (wrapper-inlined) significand expression. A leaf significand
-/// is `c * floorDiv(S * pow(E), W) [+ off]` (`c` the precision factor — `1` full,
-/// `2` the round-to-odd half-cell — and `off` the round-UP / odd `+ 1`); from it
-/// the rung supplies the window floor `m = floorDiv(...)`, its numerator
-/// `u = S * pow(E)`, the precision exponent `E` (the window is cited at `E + 1`),
-/// and `c`. Returns `None` when the significand has no floor (the carry
-/// renormalization `pow(i-1)` or the round-to-odd `n ≤ 1` constant) — those leaves
-/// borrow a sibling floor or read the format-normalization window.
-struct RfFloorInfo {
-    m: String,
-    u: String,
-    eexp: String,
-    c: u8,
-}
-
-fn rf_leaf_floor(
-    sig: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-) -> Option<RfFloorInfo> {
-    use crate::ast::{BinOp, Expr, Literal};
-    let is_one = |e: &Expr| matches!(e, Expr::Literal(Literal::Int(1)));
-    let is_two = |e: &Expr| matches!(e, Expr::Literal(Literal::Int(2)));
-    let is_floordiv = |e: &crate::ast::Spanned<Expr>| -> bool {
-        matches!(&e.node, Expr::FnCall(c, _)
-            if super::shared::expr_dotted_name(c)
-                .map(|n| rf_is_floordiv_wrapper(ctx, rf_bare_basename(&aver_name_to_lean(&n))))
-                == Some(true))
-    };
-    // Strip a trailing `+ 1` (round-up / round-to-odd).
-    let inner = match &sig.node {
-        Expr::BinOp(BinOp::Add, a, b) if is_one(&b.node) => a.as_ref(),
-        _ => sig,
-    };
-    // `c * floorDiv(...)` (c = 2) or a bare `floorDiv(...)` (c = 1).
-    let (fd, c): (&crate::ast::Spanned<Expr>, u8) = match &inner.node {
-        Expr::BinOp(BinOp::Mul, a, b) if is_two(&a.node) && is_floordiv(b) => (b.as_ref(), 2),
-        Expr::BinOp(BinOp::Mul, a, b) if is_two(&b.node) && is_floordiv(a) => (a.as_ref(), 2),
-        _ if is_floordiv(inner) => (inner, 1),
-        _ => return None,
-    };
-    let Expr::FnCall(_, fargs) = &fd.node else {
-        return None;
-    };
-    let numer = fargs.first()?;
-    // numer = S * pow(E).
-    let Expr::BinOp(BinOp::Mul, _s, powcall) = &numer.node else {
-        return None;
-    };
-    let Expr::FnCall(_, pargs) = &powcall.node else {
-        return None;
-    };
-    let eexp = rf_emit(ctx, pargs.first()?);
-    Some(RfFloorInfo {
-        m: rf_emit(ctx, fd),
-        u: rf_emit(ctx, numer),
-        eexp,
-        c,
-    })
-}
-
-/// Walk a (possibly nested) record-returning `match` — a PIECEWISE rounded value
-/// `fpValueGeneral`'s argument inlines to — and collect its leaf records'
-/// `(sigBits, exp)` fields, expanding each record's `sigBits` through
-/// [`rf_expand_sig`] (so `awaySig`'s own match contributes its arms). Returns
-/// `None` if a leaf is neither a `match` nor a `RecordCreate` (an unrecognized
-/// shape — decline rather than emit a wrong split).
-fn rf_collect_leaves(
-    expr: &crate::ast::Spanned<crate::ast::Expr>,
-    ctx: &CodegenContext,
-    out: &mut Vec<RfLeaf>,
-) -> Option<()> {
-    use crate::ast::Expr;
-    match &expr.node {
-        Expr::Match { arms, .. } => {
-            for a in arms {
-                rf_collect_leaves(&a.body, ctx, out)?;
-            }
-            Some(())
-        }
-        Expr::RecordCreate { fields, .. } => {
-            let sig = fields
-                .iter()
-                .find(|(n, _)| n == "sigBits")
-                .map(|(_, e)| e)?;
-            let exp = fields.iter().find(|(n, _)| n == "exp").map(|(_, e)| e)?;
-            let width = fields.iter().find(|(n, _)| n == "width").map(|(_, e)| e)?;
-            for s in rf_expand_sig(sig, ctx) {
-                out.push(RfLeaf {
-                    sig: s,
-                    exp: exp.clone(),
-                    width: width.clone(),
-                });
-            }
-            Some(())
-        }
-        _ => None,
-    }
-}
-
-/// The all-exponent rational strict-order rung for a PIECEWISE rounded value
-/// (Lemma 7.2.3 `away` / 7.2.4 `sticky` strict bounds): the same bound as the
-/// single-record arm, but the rounded value `fpValueGeneral(G)` inlines to a
-/// RECORD-RETURNING MATCH (the `awaySig` exact/round-up split plus the carry-to-2
-/// renormalization, or `sticky`'s round-to-odd split), not one clean record. The
-/// rung unfolds the rounding layer, `repeat' split`s the match into its leaf
-/// records, and closes EVERY leaf by one application of the generic
-/// `away_leaf` support lemma — shape-keyed on the leaf's significand `Q` and its
-/// exponent offset relative to `e_x` (`k2 = 1` no carry, `k2 = 2` the carry, whose
-/// value-uniformity is discharged by the cited pow2 homomorphism `hdbl`). The
-/// per-leaf placement `hQe` is the LINEAR fact that the effective significand sits
-/// at the floor or its successor — closed by `omega`. Name-blind, reused
-/// unchanged for `away` and `sticky`. Returns `None` (→ single-record arm) when
-/// the rounded value is not a record-returning match.
-fn emit_rational_floor_bound_matched(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Option<AutoProof> {
-    use crate::ast::Expr;
-    let pow = rf_pow2_fn(vb, law, ctx)?;
-    let inlined = inline_fn_call(&law.lhs, ctx)?;
-    let Expr::FnCall(_lt, lt_args) = &inlined.node else {
-        return None;
-    };
-    if lt_args.len() != 2 {
-        return None;
-    }
-    let Expr::FnCall(_abs, abs_args) = &lt_args[0].node else {
-        return None;
-    };
-    let err_call = abs_args.first()?;
-    let Expr::FnCall(_ps, ps_args) = &lt_args[1].node else {
-        return None;
-    };
-    let le_expr = ps_args.first()?;
-    let Expr::FnCall(_te, te_args) = &err_call.node else {
-        return None;
-    };
-    if te_args.len() != 2 {
-        return None;
-    }
-    let f_name = expr_var_name(&te_args[0].node)?.to_string();
-    let i_name = expr_var_name(&te_args[1].node)?.to_string();
-    let minus_inlined = rf_inline_fn_call(err_call, ctx)?;
-    let Expr::FnCall(_m, m_args) = &minus_inlined.node else {
-        return None;
-    };
-    if m_args.len() != 2 {
-        return None;
-    }
-    // Identify which `minus` arg is the ORIGINAL value `fpValueGeneral(F)` (a bare
-    // given) and which is the ROUNDED value `fpValueGeneral(round(F, i))`. The
-    // single-record arm assumes `(orig, rounded)`; `awayError` is `(rounded, orig)`.
-    let arg_round_call = |e: &crate::ast::Spanned<Expr>| -> Option<crate::ast::Spanned<Expr>> {
-        let Expr::FnCall(_, a) = &e.node else {
-            return None;
-        };
-        let inner = a.first()?;
-        matches!(&inner.node, Expr::FnCall(..)).then(|| inner.clone())
-    };
-    let (orig_value, rounded_call) = match (arg_round_call(&m_args[0]), arg_round_call(&m_args[1]))
-    {
-        // both are calls (round(...)): the rounded one is whichever has 2 args.
-        (Some(a0), Some(a1)) => {
-            let two = |c: &crate::ast::Spanned<Expr>| matches!(&c.node, Expr::FnCall(_, ar) if ar.len() == 2);
-            if two(&a1) {
-                (m_args[0].clone(), a1)
-            } else {
-                (m_args[1].clone(), a0)
-            }
-        }
-        (Some(a0), None) => (m_args[1].clone(), a0),
-        (None, Some(a1)) => (m_args[0].clone(), a1),
-        (None, None) => return None,
-    };
-    let fpv_f = rf_inline_fn_call(&orig_value, ctx)?;
-    let gf = rf_general_value_fields(&fpv_f, ctx, &pow)?;
-    let value_base = match &orig_value.node {
-        Expr::FnCall(c, _) => rf_bare_basename(&super::shared::expr_dotted_name(c)?).to_string(),
-        _ => return None,
-    };
-    // The rounded value must inline to a record-returning MATCH (else the
-    // single-record arm handles it).
-    let rounded_rec = rf_inline_fn_call(&rounded_call, ctx)?;
-    if !matches!(&rounded_rec.node, Expr::Match { .. }) {
-        return None;
-    }
-    let mut leaves: Vec<RfLeaf> = Vec::new();
-    rf_collect_leaves(&rounded_rec, ctx, &mut leaves)?;
-    if leaves.is_empty() {
-        return None;
-    }
-
-    let (window_thm, window_fn) = rf_window_law(vb, law, ctx)?;
-    let hom = rf_homomorphism_name(ctx, &pow)?;
-    let le = rf_emit(ctx, le_expr);
-
-    // Sub-expression strings (signed power of two kept ABSTRACT).
-    let sgnfn = gf.sgn_fn.clone();
-    let exp = gf.exp.clone();
-    let sign = gf.sign.clone();
-    let sig = gf.sig.clone();
-    let pet = format!("({sgnfn} ({exp})).top");
-    let peb = format!("({sgnfn} ({exp})).bottom");
-    let plt = format!("({sgnfn} ({le})).top");
-    let plb = format!("({sgnfn} ({le})).bottom");
-    let pe1t = format!("({sgnfn} ({exp} + 1)).top");
-    let pe1b = format!("({sgnfn} ({exp} + 1)).bottom");
-    let pw = format!("{pow} ({} - 1)", gf.width);
-    let pp = format!("{pow} ({i_name} - 1)");
-
-    // The PRIMARY floor — the first leaf bearing a floorDiv (its half/full
-    // precision window); a carry-renormalization leaf (no floor of its own) borrows
-    // it. At least one leaf must bear a floor.
-    let base = format!("{}_law_{}__rfb", aver_name_to_lean(&vb.fn_name), law.name);
-    let primary = leaves
-        .iter()
-        .find_map(|l| rf_leaf_floor(&rf_inline_wrappers(&l.sig, ctx), ctx));
-
-    // Classify each leaf's exponent offset (0 → no carry, 1 → carry) and its floor
-    // placement. Dedup by (goal-form significand, offset). Each leaf is one apply of
-    // the generic two-sided leaf lemma; name-blind, keyed on the significand shape:
-    //  * floor-bearing → its own (m, u, c) and a window cited at its precision;
-    //  * carry (no floor, exp + 1) → borrows the primary floor, k2 = 2;
-    //  * degenerate (no floor, exp + 0 — the round-to-odd n ≤ 1 constant) → m = 1,
-    //    u = S, c = 1 and the format-normalization window (1 ≤ s_x < 2).
-    let mut alts: Vec<String> = Vec::new();
-    let mut seen: std::collections::HashSet<(String, u8)> = std::collections::HashSet::new();
-    let mut has_carry = false;
-    let mut needs_degenerate = false;
-    let mut win_exps: Vec<String> = Vec::new();
-    for l in &leaves {
-        let goal_sig = rf_inline_wrappers(&l.sig, ctx);
-        let q = rf_emit(ctx, &goal_sig);
-        let le_exp = rf_emit(ctx, &l.exp);
-        let offset: u8 = if le_exp == exp {
-            0
-        } else if matches!(&l.exp.node, Expr::BinOp(crate::ast::BinOp::Add, a, b)
-            if rf_emit(ctx, a) == exp && matches!(&b.node, Expr::Literal(crate::ast::Literal::Int(1))))
-        {
-            1
-        } else {
-            return None;
-        };
-        if !seen.insert((q.clone(), offset)) {
-            continue;
-        }
-        let (av, ab, k2) = if offset == 0 {
-            (pet.clone(), peb.clone(), "1")
-        } else {
-            has_carry = true;
-            (pe1t.clone(), pe1b.clone(), "2")
-        };
-        let width_leaf = rf_emit(ctx, &l.width);
-        let p_leaf = format!("{pow} ({width_leaf} - 1)");
-        let (m, u, c, ehalf) = match rf_leaf_floor(&goal_sig, ctx) {
-            Some(fi) => {
-                if !win_exps.contains(&fi.eexp) {
-                    win_exps.push(fi.eexp.clone());
-                }
-                let e = fi.eexp.clone();
-                (fi.m, fi.u, fi.c, Some(e))
-            }
-            None if offset == 1 => {
-                let pr = primary.as_ref()?;
-                if !win_exps.contains(&pr.eexp) {
-                    win_exps.push(pr.eexp.clone());
-                }
-                (pr.m.clone(), pr.u.clone(), pr.c, Some(pr.eexp.clone()))
-            }
-            None => {
-                needs_degenerate = true;
-                (String::from("1"), sig.clone(), 1u8, None)
-            }
-        };
-        // The precision-halving discharge: `S * p = c * u` needs
-        // `pow(E+1) = 2 * pow(E)` for the half-cell (c = 2) leaf, whose `pow(i-1)`
-        // the `pow_succ_p` grind pattern (keyed on `pow(n+1)`) misses (the index is
-        // `i - 1`, not `n + 1`). Cite `pow_succ_p` at the leaf's own exponent E so
-        // grind sees the link; the `by omega` side condition (`0 ≤ E`) holds in the
-        // leaf's split branch and harmlessly fails (→ falls through) for the
-        // degenerate leaf.
-        let half = match &ehalf {
-            Some(e) => {
-                format!(" | (have hps := {base}__pow_succ_p ({e}) (by omega); grind)")
-            }
-            None => String::new(),
-        };
-        alts.push(format!(
-            "    | (apply {base}__away_leaf ({pet}) ({peb}) ({plt}) ({plb}) ({av}) ({ab}) ({pw}) ({p_leaf}) ({u}) {c} {k2} ({q}) ({sig}) ({sign}) ({m}) <;> first | assumption | omega | grind{half})"
-        ));
-    }
-
-    // The `when` conjuncts → intro pattern + the `isFp` sign conjunct.
-    let when = law.when.as_ref()?;
-    let conjs = rf_flatten_bool_and(when);
-    let sign_conj = rf_sign_conjunct(law, ctx)?;
-    let fmt_idx = conjs
-        .iter()
-        .position(|c| arith_eq(&c.node, &sign_conj.node))?;
-    let isfp_lean = match &sign_conj.node {
-        Expr::FnCall(callee, _) => aver_name_to_lean(&super::shared::expr_dotted_name(callee)?),
-        _ => return None,
-    };
-    let ncj = conjs.len();
-    let mut pat = "h_rfbp0".to_string();
-    for k in 1..ncj {
-        pat = format!("⟨{pat}, h_rfbp{k}⟩");
-    }
-    let fmtname = format!("h_rfbp{fmt_idx}");
-
-    let mut support_lines: Vec<String> =
-        super::floor_window::pow2_signed_pos_support(&base, &pow, &sgnfn)
-            .lines()
-            .map(|l| l.to_string())
-            .collect();
-    support_lines.extend(
-        super::floor_window::matched_leaf_support(&base, &pow)
-            .lines()
-            .map(|l| l.to_string()),
-    );
-
-    // The two unfold sets: the rounding layer (set1, before the split) and the
-    // value/rational layer (set2, per leaf after the split). The signed power of
-    // two stays abstract throughout; `pow`/floorDiv are already excluded.
-    let rational_prims = [
-        "absFraction",
-        "absInt",
-        "lessThan",
-        "minus",
-        "times",
-        "plus",
-        "sameValue",
-        "negate",
-    ];
-    let sgn_base = rf_bare_basename(&sgnfn).to_string();
-    let filtered = rf_filtered_defs(ctx, vb, law)
-        .into_iter()
-        .filter(|d| rf_bare_basename(d) != sgn_base)
-        .collect::<Vec<_>>();
-    let is_set2 = |d: &str| {
-        let b = rf_bare_basename(d);
-        b == value_base || rational_prims.contains(&b)
-    };
-    let set1 = filtered
-        .iter()
-        .filter(|d| !is_set2(d))
-        .cloned()
-        .collect::<Vec<_>>()
-        .join(", ");
-    let set2 = filtered
-        .iter()
-        .filter(|d| is_set2(d))
-        .cloned()
-        .collect::<Vec<_>>()
-        .join(", ");
-
-    let intro = format!("  intro {} h_when", intro_names.join(" "));
-    let mut body: Vec<String> = vec![intro, "  first".to_string(), "  | (".to_string()];
-    body.push("    simp only [Bool.and_eq_true, decide_eq_true_eq] at h_when".to_string());
-    body.push(format!("    obtain {pat} := h_when"));
-    body.push(format!("    have hpp : 0 < {pp} := {base}__pow_pos _"));
-    body.push(format!("    have hpw : 0 < {pw} := {base}__pow_pos _"));
-    body.push(format!("    have hPSeT : 0 < {pet} := {base}__sgnt_pos _"));
-    body.push(format!("    have hPSeB : 0 < {peb} := {base}__sgnb_pos _"));
-    body.push(format!("    have hPSlT : 0 < {plt} := {base}__sgnt_pos _"));
-    body.push(format!("    have hPSlB : 0 < {plb} := {base}__sgnb_pos _"));
-    if has_carry {
-        body.push(format!(
-            "    have hPSe1T : 0 < {pe1t} := {base}__sgnt_pos _"
-        ));
-        body.push(format!(
-            "    have hPSe1B : 0 < {pe1b} := {base}__sgnb_pos _"
-        ));
-    }
-    body.push(format!("    have hsign : {sign} = 1 ∨ {sign} = -1 := by unfold {isfp_lean} at {fmtname}; simp only [Bool.and_eq_true, Bool.or_eq_true, beq_iff_eq, decide_eq_true_eq] at {fmtname}; exact {fmtname}.1"));
-    body.push(format!(
-        "    have hpowi : {pow} {i_name} = 2 * {pp} := {base}__pow_of_pos {i_name} (by omega)"
-    ));
-    if has_carry {
-        body.push(format!("    have hp1 : {pow} 1 = 2 := by have hx := {base}__pow_of_pos 1 (by omega); rw [show (1:Int) - 1 = 0 by omega, {base}__pow_of_nonpos 0 (by omega)] at hx; omega"));
-    }
-    // The floor windows, one per distinct precision the leaves use: cite
-    // truncFitsWindow at the leaf's own precision E + 1 and normalize the unfolded
-    // exponent (E + 1) - 1 back to E so the floor matches the goal's
-    // floorDiv(S * pow(E), W). The away/trunc precision is i (E = i-1); the
-    // round-to-odd half-cell precision is i-1 (E = i-2).
-    for (wi, ee) in win_exps.iter().enumerate() {
-        body.push(format!(
-            "    have h_win{wi} := {window_thm} {f_name} (({ee}) + 1)"
-        ));
-        body.push(format!(
-            "    simp only [{window_fn}, Bool.and_eq_true, decide_eq_true_eq] at h_win{wi}"
-        ));
-        body.push(format!(
-            "    rw [show (({ee}) + 1) - 1 = ({ee}) by omega] at h_win{wi}"
-        ));
-        body.push(format!("    obtain ⟨hwlo{wi}, hwhi{wi}⟩ := h_win{wi}"));
-    }
-    // The format-normalization window for the degenerate (round-to-odd n ≤ 1) leaf:
-    // 1 ≤ s_x < 2, i.e. pow(width-1) ≤ S < 2 * pow(width-1), read off isFp.
-    if needs_degenerate {
-        body.push(format!("    have hnorm := {fmtname}"));
-        body.push(format!("    unfold {isfp_lean} at hnorm"));
-        body.push("    simp only [Bool.and_eq_true, Bool.or_eq_true, beq_iff_eq, decide_eq_true_eq] at hnorm".to_string());
-        body.push(format!(
-            "    have hp0 : {pow} (1 - 1) = 1 := {base}__pow_of_nonpos (1 - 1) (by omega)"
-        ));
-        body.push(format!("    have hwgt : ¬ ({} <= 0) := by", gf.width));
-        body.push("      intro hle".to_string());
-        body.push("      have hn1 := hnorm.2.1".to_string());
-        body.push("      have hn2 := hnorm.2.2".to_string());
-        body.push(format!(
-            "      rw [{base}__pow_of_nonpos ({} - 1) (by omega)] at hn1",
-            gf.width
-        ));
-        body.push(format!(
-            "      rw [{base}__pow_of_nonpos {} (by omega)] at hn2",
-            gf.width
-        ));
-        body.push("      omega".to_string());
-        body.push(format!(
-            "    have hwsucc : {pow} {} = 2 * {pow} ({} - 1) := {base}__pow_of_pos {} hwgt",
-            gf.width, gf.width, gf.width
-        ));
-        body.push(format!(
-            "    have hwlo_deg : {pow} ({} - 1) * 1 <= {sig} := by have := hnorm.2.1; omega",
-            gf.width
-        ));
-        body.push(format!(
-            "    have hwhi_deg : {sig} < {pow} ({} - 1) * (1 + 1) := by",
-            gf.width
-        ));
-        body.push(format!(
-            "      rw [show {pow} ({} - 1) * (1 + 1) = 2 * {pow} ({} - 1) by omega, ← hwsucc]",
-            gf.width, gf.width
-        ));
-        body.push("      exact hnorm.2.2".to_string());
-    }
-    // hlink: pet * plb = plt * p * peb (signed-power exponent link).
-    body.push(format!(
-        "    have hlink : {pet} * {plb} = {plt} * {pp} * {peb} := by"
-    ));
-    body.push(format!("      by_cases hE : {exp} < 0"));
-    body.push(format!("      · have hLE : {le} < 0 := by omega"));
-    body.push(format!("        have hh := {hom} ({i_name} - 1) (0 - {exp}) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"));
-    body.push(format!(
-        "        rw [show ({i_name} - 1) + (0 - {exp}) = 0 - ({le}) by omega] at hh"
-    ));
-    body.push(format!(
-        "        unfold {sgnfn}; rw [if_pos hE, if_pos hLE]; grind"
-    ));
-    body.push(format!("      · by_cases hLE : {le} < 0"));
-    body.push(format!("        · have hh := {hom} {exp} (0 - ({le})) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"));
-    body.push(format!(
-        "          rw [show {exp} + (0 - ({le})) = {i_name} - 1 by omega] at hh"
-    ));
-    body.push(format!(
-        "          unfold {sgnfn}; rw [if_neg hE, if_pos hLE]; grind"
-    ));
-    body.push(format!("        · have hh := {hom} ({le}) ({i_name} - 1) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"));
-    body.push(format!(
-        "          rw [show ({le}) + ({i_name} - 1) = {exp} by omega] at hh"
-    ));
-    body.push(format!(
-        "          unfold {sgnfn}; rw [if_neg hE, if_neg hLE]; grind"
-    ));
-    if has_carry {
-        // hdbl: pe1t * peb = 2 * pet * pe1b (the carry value-uniformity, via the
-        // same cited pow2 homomorphism with precision shift 1).
-        body.push(format!(
-            "    have hdbl : {pe1t} * {peb} = 2 * {pet} * {pe1b} := by"
-        ));
-        body.push(format!("      by_cases hE : ({exp} + 1) < 0"));
-        body.push(format!("      · have hLE : {exp} < 0 := by omega"));
-        body.push(format!("        have hh := {hom} 1 (0 - ({exp} + 1)) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"));
-        body.push(format!(
-            "        rw [show (1 : Int) + (0 - ({exp} + 1)) = 0 - {exp} by omega, hp1] at hh"
-        ));
-        body.push(format!(
-            "        unfold {sgnfn}; rw [if_pos hE, if_pos hLE]; grind"
-        ));
-        body.push(format!("      · by_cases hLE : {exp} < 0"));
-        body.push(format!("        · have hh := {hom} ({exp} + 1) (0 - {exp}) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"));
-        body.push(format!(
-            "          rw [show ({exp} + 1) + (0 - {exp}) = 1 by omega, hp1] at hh"
-        ));
-        body.push(format!(
-            "          unfold {sgnfn}; rw [if_neg hE, if_pos hLE]; grind"
-        ));
-        body.push(format!("        · have hh := {hom} {exp} 1 (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"));
-        body.push("          rw [hp1] at hh".to_string());
-        body.push(format!(
-            "          unfold {sgnfn}; rw [if_neg hE, if_neg hLE]; grind"
-        ));
-    }
-    body.push(format!("    simp only [{set1}]"));
-    body.push("    repeat' split".to_string());
-    body.push(format!(
-        "    all_goals (simp only [{set2}, decide_eq_true_eq])"
-    ));
-    body.push("    all_goals (try simp only [beq_iff_eq] at *)".to_string());
-    body.push("    all_goals (".to_string());
-    body.push("      first".to_string());
-    for a in &alts {
-        body.push(a.clone());
-    }
-    body.push("    )".to_string());
-    body.push("  )".to_string());
-
-    let floor = if super::super::tactic_ir::speculative::probing() {
-        let id = format!("{}.{}", vb.fn_name, law.name);
-        super::super::tactic_ir::speculative::record_probed(&id);
-        format!("  | (trace \"AVERSPEC_SORRY:{id}\"; sorry)")
-    } else {
-        "  | sorry".to_string()
-    };
-    body.push(floor);
-
-    Some(AutoProof {
-        support_lines,
-        body: Tactic::raw(body),
-        replaces_theorem: false,
-    })
-}
-
-/// The all-exponent rational strict-order rung for the rounding-error bound
-/// `|epsilon| < 2^(e_x - i + 1)` over the GENERAL-exponent value `fpValueGeneral`
-/// (Lemma 7.2.2's bound, faithful for the fractional floats `exp < 0` the kernel
-/// divider lives in). The error reads `minus(fpValueGeneral(F), fpValueGeneral(G))`
-/// (`G` the rounded value), so the SIGNED power of two `pow2Signed(e_x)` is the
-/// shared factor of both terms. The rung:
-///   * keeps the signed power of two an ABSTRACT atom (supplies its top/bottom
-///     positivity from `pow2_signed_pos_support`) — routing AWAY from the pow2
-///     normalizer that explodes on the squared denominators;
-///   * factors the error magnitude (sign cancels, floor remainder `r` is the
-///     nonneg core), resolves the two `absInt` matches;
-///   * CROSS-MULTIPLIES into a sign-aware INTEGER inequality and rewrites it into
-///     the multiply-by-positive shape `M * r < M * 2^(width-1)`;
-///   * derives the exponent link `2^(e_x) = 2^(e_x-i+1) * 2^(i-1)` GENERICALLY
-///     (one sign case-split, each arm CITING the power-of-two homomorphism pool
-///     law) and lets `grind` discharge the ring rewrite with it;
-///   * closes via the cited floor window (`r < 2^(width-1)`) and the
-///     multiply-by-positive rung in `aver_int_order`.
-///
-/// Returns `None` (→ legacy `fpValue` arm / bounded fallback) when the value is
-/// not the signed-power-of-two × fraction form.
-fn emit_rational_floor_bound_general(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Option<AutoProof> {
-    use crate::ast::Expr;
-    let pow = rf_pow2_fn(vb, law, ctx)?;
-    let inlined = inline_fn_call(&law.lhs, ctx)?;
-    let Expr::FnCall(_lt, lt_args) = &inlined.node else {
-        return None;
-    };
-    if lt_args.len() != 2 {
-        return None;
-    }
-    let Expr::FnCall(_abs, abs_args) = &lt_args[0].node else {
-        return None;
-    };
-    let err_call = abs_args.first()?;
-    let Expr::FnCall(_ps, ps_args) = &lt_args[1].node else {
-        return None;
-    };
-    let le_expr = ps_args.first()?;
-    let Expr::FnCall(_te, te_args) = &err_call.node else {
-        return None;
-    };
-    if te_args.len() != 2 {
-        return None;
-    }
-    let f_name = expr_var_name(&te_args[0].node)?.to_string();
-    let i_name = expr_var_name(&te_args[1].node)?.to_string();
-    // truncError ↦ minus(fpValueGeneral(F), fpValueGeneral(G)).
-    let minus_inlined = rf_inline_fn_call(err_call, ctx)?;
-    let Expr::FnCall(_m, m_args) = &minus_inlined.node else {
-        return None;
-    };
-    if m_args.len() != 2 {
-        return None;
-    }
-    let rounded_call = match &m_args[1].node {
-        Expr::FnCall(_, fv_args) => fv_args.first()?,
-        _ => return None,
-    };
-    // fpValueGeneral(F) ↦ times(SGN(exp), Fraction): read the field accessors.
-    let fpv_f = rf_inline_fn_call(&m_args[0], ctx)?;
-    let gf = rf_general_value_fields(&fpv_f, ctx, &pow)?;
-    // rounded(F, I) ↦ Fp record: the rounded significand integer (the floor q).
-    let rounded_rec = rf_inline_fn_call(rounded_call, ctx)?;
-    let Expr::RecordCreate {
-        fields: rf_fields, ..
-    } = &rounded_rec.node
-    else {
-        return None;
-    };
-    let q_spanned = rf_fields
-        .iter()
-        .find(|(n, _)| n == "sigBits")
-        .map(|(_, e)| e)?;
-    let q = rf_emit(ctx, q_spanned);
-    let le = rf_emit(ctx, le_expr);
-
-    // Cited pool laws: the floor window (strict remainder bound) and the
-    // power-of-two homomorphism (the exponent link, generically sign-split).
-    let (window_thm, window_fn) = rf_window_law(vb, law, ctx)?;
-    let hom = rf_homomorphism_name(ctx, &pow)?;
-
-    // The format predicate conjunct (its leading disjunction is the float sign).
-    let when = law.when.as_ref()?;
-    let conjs = rf_flatten_bool_and(when);
-    let sign_conj = rf_sign_conjunct(law, ctx)?;
-    let fmt_idx = conjs
-        .iter()
-        .position(|c| arith_eq(&c.node, &sign_conj.node))?;
-    let isfp_lean = match &sign_conj.node {
-        Expr::FnCall(callee, _) => aver_name_to_lean(&super::shared::expr_dotted_name(callee)?),
-        _ => return None,
-    };
-    let n = conjs.len();
-    let mut pat = "h_rfbp0".to_string();
-    for k in 1..n {
-        pat = format!("⟨{pat}, h_rfbp{k}⟩");
-    }
-    let fmtname = format!("h_rfbp{fmt_idx}");
-
-    // Cone simp set EXCLUDING the signed power of two (kept abstract) — the
-    // recursive `pow` and the floor wrapper are already excluded by the filter.
-    let sgn_base = rf_bare_basename(&gf.sgn_fn).to_string();
-    let defs = rf_filtered_defs(ctx, vb, law)
-        .into_iter()
-        .filter(|d| rf_bare_basename(d) != sgn_base)
-        .collect::<Vec<_>>()
-        .join(", ");
-
-    // Power-of-two + signed-power positivity support, scoped to a fresh prefix.
-    let base = format!("{}_law_{}__rfb", aver_name_to_lean(&vb.fn_name), law.name);
-    let support_lines: Vec<String> =
-        super::floor_window::pow2_signed_pos_support(&base, &pow, &gf.sgn_fn)
-            .lines()
-            .map(|l| l.to_string())
-            .collect();
-
-    // Derived sub-expressions (all from the cone defs, so they parse-match the
-    // unfolded goal). `pe*` are the shared signed power of two at `exp`; `pl*`
-    // the bound's signed power of two at `le`.
-    let exp = gf.exp.clone();
-    let sgn = gf.sgn_fn.clone();
-    let pet = format!("({sgn} ({exp})).top");
-    let peb = format!("({sgn} ({exp})).bottom");
-    let plt = format!("({sgn} ({le})).top");
-    let plb = format!("({sgn} ({le})).bottom");
-    let pw = format!("{pow} ({} - 1)", gf.width);
-    let pp = format!("{pow} ({i_name} - 1)");
-    let sg = &gf.sign;
-    let s = &gf.sig;
-    let r = format!("{s} * {pp} - {pw} * {q}");
-    let minustop =
-        format!("{pet} * ({sg} * {s}) * ({peb} * {pp}) - {pet} * ({sg} * {q}) * ({peb} * {pw})");
-    let t_abs = format!("{pet} * {peb} * ({r})");
-    let absbot = format!("{peb} * {pw} * ({peb} * {pp})");
-    let m = format!("{pet} * {peb} * ({absbot}) * {plb} * {plb}");
-
-    let intro = format!("  intro {} h_when", intro_names.join(" "));
-    let mut body: Vec<String> = vec![intro, "  first".to_string(), "  | (".to_string()];
-    body.push("    simp only [Bool.and_eq_true, decide_eq_true_eq] at h_when".to_string());
-    body.push(format!("    obtain {pat} := h_when"));
-    body.push(format!("    have hpp : 0 < {pp} := {base}__pow_pos _"));
-    body.push(format!("    have hpw : 0 < {pw} := {base}__pow_pos _"));
-    body.push(format!("    have hPSeT : 0 < {pet} := {base}__sgnt_pos _"));
-    body.push(format!("    have hPSeB : 0 < {peb} := {base}__sgnb_pos _"));
-    body.push(format!("    have hPSlT : 0 < {plt} := {base}__sgnt_pos _"));
-    body.push(format!("    have hPSlB : 0 < {plb} := {base}__sgnb_pos _"));
-    body.push(format!("    have hsign : {sg} = 1 ∨ {sg} = -1 := by unfold {isfp_lean} at {fmtname}; simp only [Bool.and_eq_true, Bool.or_eq_true, beq_iff_eq, decide_eq_true_eq] at {fmtname}; exact {fmtname}.1"));
-    body.push(format!("    have h_win := {window_thm} {f_name} {i_name}"));
-    body.push(format!(
-        "    simp only [{window_fn}, Bool.and_eq_true, decide_eq_true_eq] at h_win"
-    ));
-    body.push(format!(
-        "    have hr0 : 0 ≤ {r} := by have := h_win.1; omega"
-    ));
-    body.push(format!(
-        "    have hrW : {r} < {pw} := by have hexp : {pw} * ({q} + 1) = {pw} * {q} + {pw} := (by rw [Int.mul_add, Int.mul_one]); have := h_win.2; omega"
-    ));
-    // The exponent link 2^exp = 2^le * 2^(i-1), derived GENERICALLY by one
-    // exponent-sign case-split, each arm CITING the integer power-of-two
-    // homomorphism on nonnegative arguments.
-    body.push(format!(
-        "    have hlink : {pet} * {plb} = {plt} * {pp} * {peb} := by"
-    ));
-    body.push(format!("      by_cases hE : {exp} < 0"));
-    body.push(format!("      · have hLE : {le} < 0 := by omega"));
-    body.push(format!(
-        "        have hh := {hom} ({i_name} - 1) (0 - {exp}) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"
-    ));
-    body.push(format!(
-        "        rw [show ({i_name} - 1) + (0 - {exp}) = 0 - ({le}) by omega] at hh"
-    ));
-    body.push(format!(
-        "        unfold {sgn}; rw [if_pos hE, if_pos hLE]; grind"
-    ));
-    body.push(format!("      · by_cases hLE : {le} < 0"));
-    body.push(format!(
-        "        · have hh := {hom} {exp} (0 - ({le})) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"
-    ));
-    body.push(format!(
-        "          rw [show {exp} + (0 - ({le})) = {i_name} - 1 by omega] at hh"
-    ));
-    body.push(format!(
-        "          unfold {sgn}; rw [if_neg hE, if_pos hLE]; grind"
-    ));
-    body.push(format!(
-        "        · have hh := {hom} ({le}) ({i_name} - 1) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega)"
-    ));
-    body.push(format!(
-        "          rw [show ({le}) + ({i_name} - 1) = {exp} by omega] at hh"
-    ));
-    body.push(format!(
-        "          unfold {sgn}; rw [if_neg hE, if_neg hLE]; grind"
-    ));
-    body.push(format!("    simp only [{defs}, decide_eq_true_eq]"));
-    body.push(format!(
-        "    have habsN : (if {minustop} < 0 then 0 - ({minustop}) else {minustop}) = {t_abs} := by have hfact : {minustop} = {sg} * ({t_abs}) := (by grind); have hnn : 0 ≤ {t_abs} := Int.mul_nonneg (Int.mul_nonneg (Int.le_of_lt hPSeT) (Int.le_of_lt hPSeB)) hr0; rw [hfact]; rcases hsign with h | h <;> rw [h] <;> split <;> omega"
-    ));
-    body.push("    rw [habsN]".to_string());
-    body.push(format!(
-        "    have hBotpos : 0 < {absbot} := by aver_int_order"
-    ));
-    body.push(format!("    rw [if_neg (show ¬ ({absbot} < 0) by omega)]"));
-    body.push(format!(
-        "    have hL : {t_abs} * ({absbot}) * ({plb} * {plb}) = ({m}) * ({r}) := by grind"
-    ));
-    body.push(format!(
-        "    have hR : {plt} * {plb} * (({absbot}) * ({absbot})) = ({m}) * {pw} := by grind"
-    ));
-    body.push("    rw [hL, hR]".to_string());
-    body.push("    apply Int.mul_lt_mul_of_pos_left hrW".to_string());
-    body.push("    aver_int_order".to_string());
-    body.push("  )".to_string());
-
-    let floor = if super::super::tactic_ir::speculative::probing() {
-        let id = format!("{}.{}", vb.fn_name, law.name);
-        super::super::tactic_ir::speculative::record_probed(&id);
-        format!("  | (trace \"AVERSPEC_SORRY:{id}\"; sorry)")
-    } else {
-        "  | sorry".to_string()
-    };
-    body.push(floor);
-
-    Some(AutoProof {
-        support_lines,
-        body: Tactic::raw(body),
-        replaces_theorem: false,
-    })
-}
-
-/// The all-exponent rational SIGN-condition arm for the rounding-error sign
-/// law (Lemma 7.2.2's same-sign half) over the GENERAL-exponent value: the error
-/// `epsilon` has the sign of `x` — `match isNonNeg(fpValueGeneral(F)) { true =>
-/// isNonNeg(epsilon) ; false => isNonPos(epsilon) }`. The value's signed product
-/// `x.top·x.bottom` is `sign · A` with `A = (pow2Signed e_x).top·(pow2Signed
-/// e_x).bottom·sigBits·2^(width-1) > 0`, and the error's signed product is `sign ·
-/// B` with `B = (… · r · …) ≥ 0` (the same shared signed power of two and the
-/// nonneg floor remainder `r`). So whichever way the value's sign falls, the error
-/// matches it. The arm supplies the signed-power positivity (so the signed power
-/// of two stays abstract), the floor window (`r ≥ 0`) and `isFp`'s `sigBits > 0`,
-/// factors both products through `sign`, and closes by the sign case-split + the
-/// `if`/`decide` split + `omega`. Returns `None` (→ generic Sign arm) when the
-/// value is not the signed-power-of-two × fraction form.
-fn emit_rational_floor_sign_general(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Option<AutoProof> {
-    use crate::ast::Expr;
-    let pow = rf_pow2_fn(vb, law, ctx)?;
-    let inlined = inline_fn_call(&law.lhs, ctx)?;
-    let Expr::Match { subject, arms } = &inlined.node else {
-        return None;
-    };
-    if arms.len() != 2 {
-        return None;
-    }
-    // subject = isNonNeg(fpValueGeneral(F)).
-    let Expr::FnCall(_nn, nn_args) = &subject.node else {
-        return None;
-    };
-    let value_call = nn_args.first()?;
-    let fpv_f = rf_inline_fn_call(value_call, ctx)?;
-    let gf = rf_general_value_fields(&fpv_f, ctx, &pow)?;
-    // arms[0].body = isNonNeg(truncError(F, I)); get the error call.
-    let Expr::FnCall(_, a0_args) = &arms[0].body.node else {
-        return None;
-    };
-    let err_call = a0_args.first()?;
-    let Expr::FnCall(_, te_args) = &err_call.node else {
-        return None;
-    };
-    if te_args.len() != 2 {
-        return None;
-    }
-    let f_name = expr_var_name(&te_args[0].node)?.to_string();
-    let i_name = expr_var_name(&te_args[1].node)?.to_string();
-    // truncError ↦ minus(fpValueGeneral(F), fpValueGeneral(G)); read the floor q.
-    let minus_inlined = rf_inline_fn_call(err_call, ctx)?;
-    let Expr::FnCall(_m, m_args) = &minus_inlined.node else {
-        return None;
-    };
-    if m_args.len() != 2 {
-        return None;
-    }
-    let rounded_call = match &m_args[1].node {
-        Expr::FnCall(_, fv_args) => fv_args.first()?,
-        _ => return None,
-    };
-    let rounded_rec = rf_inline_fn_call(rounded_call, ctx)?;
-    let Expr::RecordCreate {
-        fields: rf_fields, ..
-    } = &rounded_rec.node
-    else {
-        return None;
-    };
-    let q_spanned = rf_fields
-        .iter()
-        .find(|(n, _)| n == "sigBits")
-        .map(|(_, e)| e)?;
-    let q = rf_emit(ctx, q_spanned);
-
-    let (window_thm, window_fn) = rf_window_law(vb, law, ctx)?;
-    let when = law.when.as_ref()?;
-    let conjs = rf_flatten_bool_and(when);
-    let sign_conj = rf_sign_conjunct(law, ctx)?;
-    let fmt_idx = conjs
-        .iter()
-        .position(|c| arith_eq(&c.node, &sign_conj.node))?;
-    let isfp_lean = match &sign_conj.node {
-        Expr::FnCall(callee, _) => aver_name_to_lean(&super::shared::expr_dotted_name(callee)?),
-        _ => return None,
-    };
-    let n = conjs.len();
-    let mut pat = "h_rfbp0".to_string();
-    for k in 1..n {
-        pat = format!("⟨{pat}, h_rfbp{k}⟩");
-    }
-    let fmtname = format!("h_rfbp{fmt_idx}");
-
-    let sgn_base = rf_bare_basename(&gf.sgn_fn).to_string();
-    let defs = rf_filtered_defs(ctx, vb, law)
-        .into_iter()
-        .filter(|d| rf_bare_basename(d) != sgn_base)
-        .collect::<Vec<_>>()
-        .join(", ");
-    let base = format!("{}_law_{}__rfs", aver_name_to_lean(&vb.fn_name), law.name);
-    let support_lines: Vec<String> =
-        super::floor_window::pow2_signed_pos_support(&base, &pow, &gf.sgn_fn)
-            .lines()
-            .map(|l| l.to_string())
-            .collect();
-
-    let exp = gf.exp.clone();
-    let pet = format!("({} ({exp})).top", gf.sgn_fn);
-    let peb = format!("({} ({exp})).bottom", gf.sgn_fn);
-    let pw = format!("{pow} ({} - 1)", gf.width);
-    let pp = format!("{pow} ({i_name} - 1)");
-    let sg = &gf.sign;
-    let s = &gf.sig;
-    let r = format!("{s} * {pp} - {pw} * {q}");
-    let minustop =
-        format!("{pet} * ({sg} * {s}) * ({peb} * {pp}) - {pet} * ({sg} * {q}) * ({peb} * {pw})");
-    let absbot = format!("{peb} * {pw} * ({peb} * {pp})");
-    let vprod = format!("{pet} * ({sg} * {s}) * ({peb} * {pw})");
-    let a_pos = format!("{pet} * {peb} * {s} * {pw}");
-    let teprod = format!("({minustop}) * ({absbot})");
-    let b_nn = format!("{pet} * {peb} * ({r}) * ({absbot})");
-
-    let intro = format!("  intro {} h_when", intro_names.join(" "));
-    let mut body: Vec<String> = vec![intro, "  first".to_string(), "  | (".to_string()];
-    body.push("    simp only [Bool.and_eq_true, decide_eq_true_eq] at h_when".to_string());
-    body.push(format!("    obtain {pat} := h_when"));
-    body.push(format!("    have hpp : 0 < {pp} := {base}__pow_pos _"));
-    body.push(format!("    have hpw : 0 < {pw} := {base}__pow_pos _"));
-    body.push(format!("    have hPSeT : 0 < {pet} := {base}__sgnt_pos _"));
-    body.push(format!("    have hPSeB : 0 < {peb} := {base}__sgnb_pos _"));
-    body.push(format!("    have hfpu := {fmtname}"));
-    body.push(format!("    unfold {isfp_lean} at hfpu"));
-    body.push(
-        "    simp only [Bool.and_eq_true, Bool.or_eq_true, beq_iff_eq, decide_eq_true_eq] at hfpu"
-            .to_string(),
-    );
-    body.push(format!("    have hsign : {sg} = 1 ∨ {sg} = -1 := hfpu.1"));
-    body.push(format!(
-        "    have hsig : 0 < {s} := by have := hfpu.2.1; omega"
-    ));
-    body.push(format!("    have h_win := {window_thm} {f_name} {i_name}"));
-    body.push(format!(
-        "    simp only [{window_fn}, Bool.and_eq_true, decide_eq_true_eq] at h_win"
-    ));
-    body.push(format!(
-        "    have hr0 : 0 ≤ {r} := by have := h_win.1; omega"
-    ));
-    body.push(format!("    simp only [{defs}, decide_eq_true_eq]"));
-    body.push(format!(
-        "    have hVfact : {vprod} = {sg} * ({a_pos}) := by grind"
-    ));
-    body.push(format!(
-        "    have hTfact : {teprod} = {sg} * ({b_nn}) := by grind"
-    ));
-    body.push(format!("    have hVA : 0 < {a_pos} := by aver_int_order"));
-    body.push(format!("    have hTB : 0 ≤ {b_nn} := by aver_int_order"));
-    body.push("    simp only [hVfact, hTfact]".to_string());
-    body.push("    rcases hsign with h | h <;> simp only [h] <;> split <;> simp only [decide_eq_true_eq] <;> omega".to_string());
-    body.push("  )".to_string());
-    let floor = if super::super::tactic_ir::speculative::probing() {
-        let id = format!("{}.{}", vb.fn_name, law.name);
-        super::super::tactic_ir::speculative::record_probed(&id);
-        format!("  | (trace \"AVERSPEC_SORRY:{id}\"; sorry)")
-    } else {
-        "  | sorry".to_string()
-    };
-    body.push(floor);
-
-    Some(AutoProof {
-        support_lines,
-        body: Tactic::raw(body),
-        replaces_theorem: false,
-    })
-}
-
-fn emit_rational_floor_bound(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Option<AutoProof> {
-    use crate::ast::Expr;
-    // A PIECEWISE rounded value (record-returning match — `away`/`sticky`) takes
-    // the matched-leaf rung; a single clean record (`trunc`) takes the
-    // single-record general rung; the legacy clamped value keeps the arm below.
-    if let Some(p) = emit_rational_floor_bound_matched(vb, law, ctx, intro_names) {
-        return Some(p);
-    }
-    // The general-exponent value (fpValueGeneral) takes the signed strict-order
-    // rung; the legacy clamped value (fpValue) keeps the arm below.
-    if let Some(p) = emit_rational_floor_bound_general(vb, law, ctx, intro_names) {
-        return Some(p);
-    }
-    let pow = rf_pow2_fn(vb, law, ctx)?;
-    // Walk the recognized shape:
-    //   lessThan(absFraction(truncError(F, I)), pow2Signed(LE))
-    //   truncError(F, I)  ↦  minus(fpValue(F), fpValue(G)),  G = rounded(F, I)
-    let inlined = inline_fn_call(&law.lhs, ctx)?;
-    let Expr::FnCall(_lt, lt_args) = &inlined.node else {
-        return None;
-    };
-    if lt_args.len() != 2 {
-        return None;
-    }
-    let Expr::FnCall(_abs, abs_args) = &lt_args[0].node else {
-        return None;
-    };
-    let err_call = abs_args.first()?;
-    let Expr::FnCall(_ps, ps_args) = &lt_args[1].node else {
-        return None;
-    };
-    let le_expr = ps_args.first()?;
-    // truncError(F, I): grab the float / precision givens directly.
-    let Expr::FnCall(_te, te_args) = &err_call.node else {
-        return None;
-    };
-    if te_args.len() != 2 {
-        return None;
-    }
-    let f_name = expr_var_name(&te_args[0].node)?.to_string();
-    let i_name = expr_var_name(&te_args[1].node)?.to_string();
-    // truncError ↦ minus(fpValue(F), fpValue(G)).
-    let minus_inlined = rf_inline_fn_call(err_call, ctx)?;
-    let Expr::FnCall(_m, m_args) = &minus_inlined.node else {
-        return None;
-    };
-    if m_args.len() != 2 {
-        return None;
-    }
-    let rounded_call = match &m_args[1].node {
-        Expr::FnCall(_, fv_args) => fv_args.first()?,
-        _ => return None,
-    };
-    // fpValue(F) ↦ Fraction record: read the field accessors off it.
-    let fpv_f = rf_inline_fn_call(&m_args[0], ctx)?;
-    let fields = rf_fpvalue_fields(&fpv_f, ctx)?;
-    // rounded(F, I) ↦ Fp record: the rounded significand integer (the floor q).
-    let rounded_rec = rf_inline_fn_call(rounded_call, ctx)?;
-    let Expr::RecordCreate {
-        fields: rf_fields, ..
-    } = &rounded_rec.node
-    else {
-        return None;
-    };
-    let q_spanned = rf_fields
-        .iter()
-        .find(|(n, _)| n == "sigBits")
-        .map(|(_, e)| e)?;
-    let q = rf_emit(ctx, q_spanned);
-    let le = rf_emit(ctx, le_expr);
-
-    // The cited pool laws.
-    let (window_thm, window_fn) = rf_window_law(vb, law, ctx)?;
-    let hom = rf_homomorphism_name(ctx, &pow)?;
-    let defs = rf_filtered_defs(ctx, vb, law).join(", ");
-
-    // The format predicate conjunct (its leading disjunction is the float sign).
-    let when = law.when.as_ref()?;
-    let conjs = rf_flatten_bool_and(when);
-    let sign_conj = rf_sign_conjunct(law, ctx)?;
-    let fmt_idx = conjs
-        .iter()
-        .position(|c| arith_eq(&c.node, &sign_conj.node))?;
-    let isfp_lean = match &sign_conj.node {
-        Expr::FnCall(callee, _) => aver_name_to_lean(&super::shared::expr_dotted_name(callee)?),
-        _ => return None,
-    };
-    let n = conjs.len();
-    let mut pat = "h_rfbp0".to_string();
-    for k in 1..n {
-        pat = format!("⟨{pat}, h_rfbp{k}⟩");
-    }
-    let fmtname = format!("h_rfbp{fmt_idx}");
-
-    // The power-of-two positivity support, scoped to a fresh per-law prefix.
-    let base = format!("{}_law_{}__rfb", aver_name_to_lean(&vb.fn_name), law.name);
-    let support_lines: Vec<String> = super::floor_window::pow_pos_support(&base, &pow)
-        .lines()
-        .map(|l| l.to_string())
-        .collect();
-
-    // Derived sub-expression strings (all from the cone defs, so they parse-match
-    // the unfolded goal).
-    let pe = format!("{pow} ({})", fields.exp);
-    let pw = format!("{pow} ({} - 1)", fields.width);
-    let pp = format!("{pow} ({i_name} - 1)");
-    let pr = format!("{pow} (0 - ({le}))");
-    let ple = format!("{pow} ({le})");
-    let sgn = &fields.sign;
-    let sig = &fields.sig;
-    let exa = &fields.exp;
-    let r = format!("{sig} * {pp} - {pw} * {q}");
-    let num = format!("{sgn} * {sig} * {pe} * {pp} - {sgn} * {q} * {pe} * {pw}");
-
-    let intro = format!("  intro {} h_when", intro_names.join(" "));
-    let mut body: Vec<String> = vec![intro, "  first".to_string(), "  | (".to_string()];
-    body.push("    simp only [Bool.and_eq_true, decide_eq_true_eq] at h_when".to_string());
-    body.push(format!("    obtain {pat} := h_when"));
-    body.push(format!("    have hP : 0 < {pp} := {base}__pow_pos _"));
-    body.push(format!("    have hW : 0 < {pw} := {base}__pow_pos _"));
-    body.push(format!("    have hpe : 0 < {pe} := {base}__pow_pos _"));
-    body.push(format!(
-        "    have hWP : 0 < {pw} * {pp} := Int.mul_pos hW hP"
-    ));
-    body.push(format!("    have h_win := {window_thm} {f_name} {i_name}"));
-    body.push(format!(
-        "    simp only [{window_fn}, Bool.and_eq_true, decide_eq_true_eq] at h_win"
-    ));
-    body.push(format!(
-        "    have hr0 : 0 ≤ {r} := by have := h_win.1; omega"
-    ));
-    body.push(format!(
-        "    have hrW : {r} < {pw} := by have hexp : {pw} * ({q} + 1) = {pw} * {q} + {pw} := (by rw [Int.mul_add, Int.mul_one]); have := h_win.2; omega"
-    ));
-    body.push(format!("    have hsign : {sgn} = 1 ∨ {sgn} = -1 := by unfold {isfp_lean} at {fmtname}; simp only [Bool.and_eq_true, Bool.or_eq_true, beq_iff_eq, decide_eq_true_eq] at {fmtname}; exact {fmtname}.1"));
-    body.push(format!("    simp only [{defs}, decide_eq_true_eq]"));
-    body.push(format!(
-        "    have habsN : (if {num} < 0 then 0 - ({num}) else {num}) = {pe} * ({r}) := by have hfact : {num} = {sgn} * ({pe} * ({r})) := (by grind); have hnn : 0 ≤ {pe} * ({r}) := Int.mul_nonneg (by omega) hr0; rw [hfact]; rcases hsign with h | h <;> rw [h] <;> split <;> omega"
-    ));
-    body.push("    rw [habsN]".to_string());
-    body.push(format!(
-        "    rw [if_neg (show ¬ ({pw} * {pp} < 0) by omega)]"
-    ));
-    body.push(format!("    by_cases hk : {le} < 0"));
-    // k < 0 branch
-    body.push("    · rw [if_pos hk]; dsimp only".to_string());
-    body.push(format!("      have hRb : 0 < {pr} := {base}__pow_pos _"));
-    body.push(format!(
-        "      have hlink : {pe} * {pr} = {pp} := by have h := {hom} ({exa}) (0 - ({le})) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega); rw [show {exa} + (0 - ({le})) = {i_name} - 1 by omega] at h; omega"
-    ));
-    body.push(format!(
-        "      have lhs_eq : {pe} * ({r}) * ({pw} * {pp}) * ({pr} * {pr}) = {pe} * ({pw} * {pp}) * ({pr} * {pr}) * ({r}) := by grind"
-    ));
-    body.push(format!(
-        "      have rhs_eq : 1 * {pr} * (({pw} * {pp}) * ({pw} * {pp})) = {pe} * ({pw} * {pp}) * ({pr} * {pr}) * {pw} := by rw [← hlink]; grind"
-    ));
-    body.push("      rw [lhs_eq, rhs_eq]; aver_int_order".to_string());
-    // k >= 0 branch
-    body.push("    · rw [if_neg hk]; dsimp only".to_string());
-    body.push(format!(
-        "      have hlink : {ple} * {pp} = {pe} := by have h := {hom} ({le}) ({i_name} - 1) (by simp only [Bool.and_eq_true, ge_iff_le, decide_eq_true_eq]; omega); rw [show ({le}) + ({i_name} - 1) = {exa} by omega] at h; omega"
-    ));
-    body.push(format!(
-        "      have lhs_eq : {pe} * ({r}) * ({pw} * {pp}) * (1 * 1) = {pe} * ({pw} * {pp}) * ({r}) := by grind"
-    ));
-    body.push(format!(
-        "      have rhs_eq : {ple} * 1 * (({pw} * {pp}) * ({pw} * {pp})) = {pe} * ({pw} * {pp}) * {pw} := by rw [← hlink]; grind"
-    ));
-    body.push("      rw [lhs_eq, rhs_eq]; aver_int_order".to_string());
-    body.push("  )".to_string());
-
-    let floor = if super::super::tactic_ir::speculative::probing() {
-        let id = format!("{}.{}", vb.fn_name, law.name);
-        super::super::tactic_ir::speculative::record_probed(&id);
-        format!("  | (trace \"AVERSPEC_SORRY:{id}\"; sorry)")
-    } else {
-        "  | sorry".to_string()
-    };
-    body.push(floor);
-
-    Some(AutoProof {
-        support_lines,
-        body: Tactic::raw(body),
-        replaces_theorem: false,
-    })
-}
-
-/// The rational-over-floor sign/magnitude keystone arm. Returns the full
-/// auto-proof (support stack + body) or `None` to fall through to the generic
-/// keystone emission.
-fn emit_rational_floor_family(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Option<AutoProof> {
-    let shape = rational_floor_shape(vb, law, ctx)?;
-    // The rational-order truncation-error bound (Lemma 7.2.2) has its own
-    // dedicated arm: factor + sign-split + cite-window + cite-homomorphism +
-    // multiply-by-positive, none of which the NonnegPos / Sign skeletons cover.
-    if shape == RationalFloorShape::Bound {
-        return emit_rational_floor_bound(vb, law, ctx, intro_names);
-    }
-    // The general-exponent same-sign condition (Lemma 7.2.2 sign half over
-    // fpValueGeneral) takes its own signed-power-of-two arm; the legacy clamped
-    // value keeps the generic Sign skeleton below.
-    if shape == RationalFloorShape::Sign
-        && let Some(p) = emit_rational_floor_sign_general(vb, law, ctx, intro_names)
-    {
-        return Some(p);
-    }
-    let pow = rf_pow2_fn(vb, law, ctx)?;
-    let citations = rf_citations(vb, law, ctx, intro_names);
-    let defs = rf_filtered_defs(ctx, vb, law).join(", ");
-    let intro = format!("  intro {} h_when", intro_names.join(" "));
-
-    // Citation `have`s (shared by both shapes): bring each cited law's order
-    // content into context, unfolded to its Int facts.
-    let mut steps: Vec<String> = Vec::new();
-    for c in &citations {
-        steps.push(format!("have {} := {}", c.have_name, c.apply));
-        steps.push(format!("simp only [{}] at {}", c.simp_set, c.have_name));
-    }
-
-    let mut support_lines: Vec<String> = Vec::new();
-    let arm: String = match shape {
-        RationalFloorShape::NonnegPos => {
-            // Power-of-two positivity at the record-field exponents.
-            let base = format!("{}_law_{}__rfpp", aver_name_to_lean(&vb.fn_name), law.name);
-            support_lines = super::floor_window::pow_pos_support(&base, &pow)
-                .lines()
-                .map(|l| l.to_string())
-                .collect();
-            let mut idx = 0usize;
-            let mut push_pow = |steps: &mut Vec<String>, expr: String| {
-                steps.push(format!("have h_rfpp{idx} := {base}__pow_pos ({expr})"));
-                idx += 1;
-            };
-            for g in &law.givens {
-                let gl = aver_name_to_lean(&g.name);
-                if g.type_name.trim() == "Int" {
-                    push_pow(&mut steps, gl.clone());
-                    push_pow(&mut steps, format!("{gl} - 1"));
-                } else {
-                    for fld in rf_record_int_fields(ctx, &g.type_name) {
-                        push_pow(&mut steps, format!("{gl}.{fld}"));
-                        push_pow(&mut steps, format!("{gl}.{fld} - 1"));
-                    }
-                }
-            }
-            let close = format!(
-                "simp only [{defs}, Bool.and_eq_true, Bool.or_eq_true, ge_iff_le, gt_iff_lt, decide_eq_true_eq, beq_iff_eq] at h_when ⊢ <;> aver_int_order"
-            );
-            steps.push(close);
-            format!("  | ({})", steps.join("; "))
-        }
-        RationalFloorShape::Sign => {
-            // The sign/format predicate conjunct (`isFp`): `rcases` its leading
-            // disjunction (the float sign) after unfolding it.
-            let sign_conj = rf_sign_conjunct(law, ctx)?;
-            let when = law.when.as_ref()?;
-            let prem = rf_premise_proof(when, sign_conj)?;
-            steps.push(format!("have h_rfsign := {prem}"));
-            steps.push(format!(
-                "simp only [{defs}, Bool.and_eq_true, Bool.or_eq_true, decide_eq_true_eq, beq_iff_eq] at h_rfsign"
-            ));
-            steps.push(format!("simp only [{defs}]"));
-            steps.push("rcases h_rfsign.1 with hs | hs <;> simp only [hs] <;> grind".to_string());
-            format!("  | ({})", steps.join("; "))
-        }
-        // Handled by the dedicated arm above (early return); never reached here.
-        RationalFloorShape::Bound => {
-            unreachable!("Bound shape handled by emit_rational_floor_bound")
-        }
-    };
-
-    let floor = if super::super::tactic_ir::speculative::probing() {
-        let id = format!("{}.{}", vb.fn_name, law.name);
-        super::super::tactic_ir::speculative::record_probed(&id);
-        format!("  | (trace \"AVERSPEC_SORRY:{id}\"; sorry)")
-    } else {
-        "  | sorry".to_string()
-    };
-
-    Some(AutoProof {
-        support_lines,
-        body: Tactic::raw(vec![intro, "  first".to_string(), arm, floor]),
-        replaces_theorem: false,
-    })
-}
-
-/// Emit the keystone proof: `intro <givens> h_when; first | (simp only [<cone>,
-/// <bridges>] at h_when ⊢ <;> grind [<cone>, <pool law names>]) | <floor>`. The
-/// `simp` unfolds the cone and bridges the Bool comparison / splits a
-/// conjunctive guard; `grind` then composes the laws-as-lemmas pool (cone defs
-/// AND the earlier-law theorem names) to close the residual. The `first | … |
-/// sorry` floor keeps credit fail-closed; under the speculative probe the floor
-/// carries the `AVERSPEC_SORRY:<fn.law>` trace so a non-closing portfolio is
-/// observable and the law falls back to its bounded statement.
-pub(in crate::codegen::lean) fn emit_pool_composition_generic_law(
-    vb: &VerifyBlock,
-    law: &VerifyLaw,
-    ctx: &CodegenContext,
-    intro_names: &[String],
-) -> Option<AutoProof> {
-    if !recognize_pool_composition_generic(vb, law, ctx) {
-        return None;
-    }
-    // The rational-over-floor sign/magnitude family (Lemma 7.2.2 sign half) takes
-    // a dedicated cite-window + pow-positivity + `aver_int_order` / sign-split
-    // arm, and deliberately bypasses the `pow2` homomorphism normalizer below
-    // (it explodes on the squared-denominator rational goals). Every other
-    // keystone law declines `rational_floor_shape` and keeps the emission below.
-    if let Some(proof) = emit_rational_floor_family(vb, law, ctx, intro_names) {
-        return Some(proof);
-    }
-    // `simp only [<cone>]` unfolds the cone — but RECURSIVE cone fns (`pow2`)
-    // must be EXCLUDED: `simp only [pow2]` errors trying to reduce the recursive
-    // match on a symbolic argument, failing the whole arm. They stay abstract
-    // and the pool law (the `pow2` homomorphism) reasons about them instead.
-    // `law_simp_defs` emits fully-qualified `_root_.<name>` simp targets, but
-    // `recursive_pure_fn_names` (via `aver_name_to_lean`) yields bare names. Strip
-    // the `_root_.` namespace prefix on BOTH sides before comparing, or the
-    // recursive filter silently misses (`"_root_.pow2" != "pow2"`) and a
-    // recursive cone fn leaks into `simp only`, breaking simp on its symbolic
-    // recursive match (the measured `e2` failure mode).
-    // Compare on the BARE basename. `recursive_pure_fn_names` yields bare names
-    // (`pow2`), but `law_simp_defs` QUALIFIES a dep-module fn (`Domain.Fprep.pow2`,
-    // and entry fns carry a `_root_.` prefix). Stripping only `_root_.` left a dep
-    // recursive fn (`Domain.Fprep.pow2`) unmatched against the bare `pow2`, so it
-    // leaked into `simp only [...]` — and `simp only` on a recursive fn's symbolic
-    // match ERRORS, failing the whole keystone arm (the cross-module miss that
-    // sent every dep-cone conditional law to its bounded fallback). Drop the
-    // `_root_.` prefix AND the module path on both sides so a dep recursive cone
-    // fn is excluded exactly like an entry one.
-    fn bare_basename(n: &str) -> &str {
-        let n = n.strip_prefix("_root_.").unwrap_or(n);
-        n.rsplit('.').next().unwrap_or(n)
-    }
-    let mut abstract_fns: std::collections::HashSet<String> = super::recursive_pure_fn_names(ctx)
-        .iter()
-        .map(|n| bare_basename(&aver_name_to_lean(n)).to_string())
-        .collect();
-    // Also keep FOLDED any cone fn that a CITED pool law is keyed on: that
-    // law is a rewrite ABOUT the fn (the floorDiv exact-cancel is about
-    // `floorDiv`), so it can only e-match while the fn is still abstract —
-    // the same reason a recursive cone fn (the `pow2` homomorphism's `pow2`)
-    // is excluded. Without this the `simp only` would unfold `floorDiv` to its
-    // `Result.withDefault`/`Int.div` body and the cancel citation would find
-    // nothing to fire on.
-    for subj in keystone_pool_subject_fns(vb, law, ctx) {
-        abstract_fns.insert(bare_basename(&subj).to_string());
-    }
-    // SIGNED power-of-two homomorphism: when an earlier law proves the
-    // homomorphism of a `Fraction`-valued signed `2^k` cone fn `F` and the goal
-    // applies `F` to a sum of exponents, keep `F` FOLDED (like a recursive cone
-    // fn) so the cited cross-multiply canonicalizer can e-match its `F(m+n)` /
-    // `F(m+n+1)` terms — unfolding `F` would bury the float-product `if` inside
-    // `F`'s sign match and explode the split. Name-blind via the recognizer.
-    let pow2signed = keystone_pow2signed_cite(vb, law, ctx);
-    if let Some(cite) = &pow2signed {
-        abstract_fns.insert(bare_basename(&cite.sgn).to_string());
-    }
-    let defs: Vec<String> = law_simp_defs(ctx, vb, law)
-        .into_iter()
-        .filter(|d| !abstract_fns.contains(bare_basename(d)))
-        .collect();
-    let defs_csv = defs.join(", ");
-    // `grind` gets the laws-as-lemmas pool names (not the cone defs): the `simp`
-    // already unfolds the non-recursive cone, and feeding grind the recursive defs
-    // risks it re-unfolding into a loop. Two pools, both fed to the same `grind`:
-    //   * EQUATIONAL — earlier laws whose subject fn is in the cone (e.g. a `pow2`
-    //     homomorphism). grind e-matches each, instantiates it, discharges its
-    //     premise from context, and closes the residual ring/linear identity.
-    //   * ORDER-BRIDGE — each cited inequality law, restated in Prop form with a
-    //     `grind_pattern` over its two comparison sides (emitted as `support_lines`
-    //     BEFORE this theorem). The pattern fires when grind sees those sides in
-    //     the goal/premises, instantiating the order fact (e.g. `a*c ≤ b*c`) that
-    //     omega/grind then chains with a premise (`b*c ≤ m`) to the conclusion.
-    // Drop self-registering (`grind_pattern`-annotated, so already `@[grind]`)
-    // cancel laws from the explicit hint list — `grind` picks them up globally,
-    // and re-listing them only warns. They stay folded via `abstract_fns` above
-    // so their floor term survives for the pattern to e-match.
-    let self_registering = keystone_self_registering_pool_names(vb, law, ctx);
-    let pool_names: Vec<String> = keystone_pool_names(vb, law, ctx)
-        .into_iter()
-        .filter(|n| !self_registering.contains(n))
-        .collect();
-    let order_citations = keystone_order_bridge_citations(vb, law, ctx);
-    let mut hints: Vec<String> = pool_names;
-    hints.extend(order_citations.iter().map(|c| c.name.clone()));
-    let mut support_lines: Vec<String> = Vec::new();
-    for citation in &order_citations {
-        support_lines.extend(citation.support.iter().cloned());
-    }
-
-    // GENERIC pow2-of-nonneg-linear-form normalizer. When the cited pool
-    // reasons about a power-of-two fn (any `is_pow2_shape` fn, detected via the
-    // cited law's `FloorDivWindow` figure — name-blind), prepend its normalizer
-    // support stack (the homomorphism keyed on the PRODUCT side + the successor
-    // step, both with `grind_pattern`s) and add those lemmas to the grind hints.
-    // This is what lets `grind` canonicalize `pow` of a REARRANGED / shifted
-    // exponent — `pow (w_a + w_b - 2)` unifies with `pow (w_a-1) * pow (w_b-1)`
-    // by the product pattern + linear-arith congruence, and an odd offset peels
-    // via the successor pattern — closing the homomorphism COMPOSITION (e.g. the
-    // folklore `fpMulValue`) that the default LHS (`pow (m+n)`) e-match misses on
-    // a subtraction. Scoped to a fresh per-law prefix: no global pattern clash.
-    let pow2_normalizer = keystone_pow2_fn(vb, law, ctx).map(|pow| {
-        let nlbase = format!(
-            "{}_law_{}__pow2lf",
-            aver_name_to_lean(&vb.fn_name),
-            law.name
-        );
-        let support = super::floor_window::pow2_linear_form_normalizer_support(&nlbase, &pow);
-        (nlbase, support)
-    });
-    if let Some((nlbase, support)) = &pow2_normalizer {
-        support_lines.extend(support.lines().map(|l| l.to_string()));
-        hints.push(format!("{nlbase}__pow_add"));
-        hints.push(format!("{nlbase}__pow_succ"));
-    }
-    // GENERIC signed-power-of-two homomorphism normalizer (the `Fraction`-level
-    // companion of the `pow2lf` integer normalizer above). Emits a self-contained
-    // support stack about the signed fn `F` — its positivity, `F(1) = 2/1`, and
-    // the cross-multiplied homomorphism CITED from the proven Aver law, keyed on
-    // the SUM side `F(m+n)` / `F(m+n+1)` — so `grind` canonicalizes `F` of a
-    // summed/shifted exponent the way the integer normalizer canonicalizes `pow`
-    // of a summed/shifted one. Scoped to a fresh per-law prefix.
-    if let Some(cite) = &pow2signed {
-        let sgnbase = format!(
-            "{}_law_{}__pow2sgn",
-            aver_name_to_lean(&vb.fn_name),
-            law.name
-        );
-        let support = super::floor_window::pow2_signed_homomorphism_normalizer_support(
-            &sgnbase,
-            &cite.pow,
-            &cite.sgn,
-            &cite.hom_fn,
-            &cite.cited,
-        );
-        support_lines.extend(support.lines().map(|l| l.to_string()));
-        hints.push(format!("{sgnbase}__sgn_add"));
-        hints.push(format!("{sgnbase}__sgn_add_succ"));
-    }
-    let grind_hints = hints.join(", ");
-    // With no citable pool and no normalizer the hint list is empty; emit a bare
-    // `grind` (not `grind []`) for the flat definitional close.
-    let grind_call = if grind_hints.is_empty() {
-        "grind".to_string()
-    } else {
-        format!("grind [{grind_hints}]")
-    };
-
-    // A non-conditional law (the unconditional signed pow2 homomorphism) has no
-    // premise to introduce, so `intro` binds only the givens and `simp` rewrites
-    // the goal alone (`at ⊢`); a `when`-law also introduces and simplifies the
-    // premise hypothesis (`at h_when ⊢`).
-    let (intro, simp_at) = if law.when.is_some() {
-        (
-            format!("  intro {} h_when", intro_names.join(" ")),
-            "at h_when ⊢",
-        )
-    } else {
-        (format!("  intro {}", intro_names.join(" ")), "at ⊢")
-    };
-    // Bridge set matches the de-risked `e1`: `Bool.and_eq_true` splits the `&&`
-    // guard, `decide_eq_true_eq` peels `decide p = true` to `p`. The earlier
-    // `ge_iff_le, gt_iff_lt` are dropped — they are unnecessary for this shape
-    // (`grind` orders relations itself) and `e1` closes without them.
-    //
-    // The signed-power-of-two homomorphism path adds `beq_iff_eq`: with `F`
-    // FOLDED the goal stays a `Fraction` `sameValue` (a `==`), and turning it into
-    // a Prop `=` lets `grind`'s commutative-ring solver consume the cited
-    // cross-multiply facts (it rejects a bare `decide (… == …)` wrapper). Added
-    // ONLY on the signed path — the integer-only pow2 laws close without it, so
-    // their emission stays byte-identical.
-    let bridge_extra = if pow2signed.is_some() {
-        ", beq_iff_eq"
-    } else {
-        ""
-    };
-    let close = if pow2_normalizer.is_some() {
-        // A multiplication-value law (`fpMul`) normalizes its significand with
-        // an `if`-branch (shift 0 vs 1); `repeat' split` discharges that branch
-        // before `grind`, and the `first | (repeat' split <;> …) | …` covers both
-        // the branching and the flat (no-`if`) shapes uniformly — the bare `grind`
-        // arm is the path the scaling / product-exponent laws (no `if`) take.
-        // `repeat'` (not a single `split`) is load-bearing for the SIGNED pow2
-        // homomorphism: its goal carries THREE independent `pow2Signed` sign
-        // matches (for `m`, `n`, and `m+n`), and `grind` does not case-split the
-        // residual matches itself — every sign arm must be peeled before `grind`
-        // sees a product of `pow2` of nonneg atoms the normalizer can canonicalize.
-        // For a single-match law `repeat' split` peels exactly the one match, so
-        // the fpMul / fpScale / product-exponent laws are unaffected.
-        format!(
-            "  | (simp only [{defs_csv}, Bool.and_eq_true, decide_eq_true_eq{bridge_extra}] {simp_at} <;> (first | ((repeat' split) <;> {grind_call}) | {grind_call}))"
-        )
-    } else {
-        format!(
-            "  | (simp only [{defs_csv}, Bool.and_eq_true, decide_eq_true_eq{bridge_extra}] {simp_at} <;> {grind_call})"
-        )
-    };
-    let floor = if super::super::tactic_ir::speculative::probing() {
-        let id = format!("{}.{}", vb.fn_name, law.name);
-        super::super::tactic_ir::speculative::record_probed(&id);
-        format!("  | (trace \"AVERSPEC_SORRY:{id}\"; sorry)")
-    } else {
-        "  | sorry".to_string()
-    };
-    Some(AutoProof {
-        support_lines,
-        body: Tactic::raw(vec![intro, "  first".to_string(), close, floor]),
         replaces_theorem: false,
     })
 }
